refactor(wiki): extract route orchestration into src/views/wiki/WikiPage.tsx
page.tsx is now a one-line re-export; WikiPage, Explorer, NewDocView,
WikiTombstone, WikiUnknownLink, SortBar, and Row move to the new views module.
