+++ frontend/src/app/app/wiki/[[...slug]]/page.tsx
@@ -1,2962 +1,1 @@
-"use client";
-
-import { useParams, useRouter, useSearchParams } from "next/navigation";
-import {
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-  type FormEvent,
-} from "react";
-import { createPortal } from "react-dom";
-import useSWR from "swr";
-import {
-  Button,
-  EmptyMessageCard,
-  LineItemButton,
-  MessageCard,
-  OpenButton,
-  Popover,
-  PopoverMenu,
-  SelectButton,
-} from "@onyx-ai/opal/components";
-import {
-  SvgAlertTriangle,
-  SvgBubbleText,
-  SvgChevronLeft,
-  SvgChevronRight,
-  SvgDocFile,
-  SvgEdit,
-  SvgExternalLink,
-  SvgFolder,
-  SvgFolderPlus,
-  SvgHistory,
-  SvgPlus,
-  SvgShare,
-  SvgShield,
-  SvgSparkle,
-  SvgTrash,
-  SvgWorkflow,
-} from "@onyx-ai/opal/icons";
-import { useConfirm } from "@/components/common/ConfirmDialog";
-import { LoadingSpinner } from "@/components/common/LoadingSpinner";
-import { TriggerPanel } from "@/components/triggers/TriggerPanel";
-import { TriggersSidePanel } from "@/components/wiki/TriggersSidePanel";
-import { useLeftPanel } from "@/providers/LeftPanelProvider";
-import { deleteTrigger, useTriggers, type Trigger } from "@/lib/triggers";
-import { DiffView } from "@/components/wiki/DiffView";
-import { HistoryPanel } from "@/components/wiki/HistoryPanel";
-import { RunAgentPanel } from "@/components/wiki/RunAgentPanel";
-import { WikiHome } from "@/components/wiki/WikiHome";
-import { craftFailureMessage } from "@/lib/craft";
-import {
-  closeSession,
-  useAgentSessions,
-  type AgentSessionSummary,
-} from "@/lib/launchers";
-import { ShareDialog } from "@/components/wiki/ShareDialog";
-import { CommentsPanel } from "@/components/wiki/CommentsPanel";
-import { UpdateHealthBanner } from "@/components/wiki/UpdateHealthBanner";
-import { UpdatePolicyPanel } from "@/components/wiki/UpdatePolicyPanel";
-import { apiFetch, ApiError } from "@/lib/api";
-import {
-  isDocId,
-  wikiHref,
-  wikiPath,
-  resolveDocId,
-  resolveIds,
-  revalidateWiki,
-} from "@/lib/wikiHref";
-import { formatRelative } from "@/lib/format";
-import { getDeletedTombstone, restoreTrashed } from "@/lib/trash";
-import { listComments } from "@/lib/comments";
-import {
-  paintCommentHighlights,
-  scrollCommentIntoView,
-  selectionToAnchor,
-  type CommentDraft,
-} from "@/lib/fileview/commentAnchor";
-import { DocTitle, MarkdownRenderer } from "@/lib/fileview/components";
-import { pageTitle } from "@/lib/fileview/utils";
-import { useAuth, useRequireAuth } from "@/lib/auth";
-import { Coeditor } from "@/lib/coeditor/components";
-import type { CoeditParticipant } from "@/lib/coeditor/types";
-import { useCoeditSession } from "@/lib/coeditor/hooks";
-import {
-  useAgentsBarHost,
-  useHeaderActionsHost,
-  useRightPanelHost,
-} from "@/providers/WikiHeaderActionsProvider";
-import { useDrafting } from "@/lib/drafting";
-import { rememberWikiPath } from "@/lib/lastViewed";
-import { recordRecentDoc } from "@/lib/recents";
-import {
-  getDraftState,
-  getTemplate,
-  listTemplateSummaries,
-  setDraftTemplate,
-  type DocumentTemplateSummary,
-} from "@/lib/templates";
-import { absoluteTime, relativeTime } from "@/lib/time";
-import { useIsMobile } from "@/lib/viewport";
-import {
-  AI_DRAFT_KEY,
-  type CommitInfo,
-  fetchFileDiff,
-  fetchFileHistory,
-  type FileDiffResponse,
-} from "@/lib/wiki";
-import type {
-  CommentThreadView,
-  DocumentActivity,
-  DocumentActivityResponse,
-} from "@/types";
-
-interface DocEntry {
-  path: string;
-  updated_at: string;
-}
-
-interface ListResponse {
-  entries: DocEntry[];
-}
-
-interface FileResponse {
-  path: string;
-  body: string;
-  ref?: string;
-  head_sha?: string | null;
-}
-
-export default function WikiRoute() {
-  const { user, loading } = useRequireAuth();
-  const isMobile = useIsMobile();
-  const router = useRouter();
-  const params = useParams<{ slug?: string[] }>();
-  const searchParams = useSearchParams();
-  const isNewMode = searchParams?.get("new") === "1";
-  const rawSlugParts = (params?.slug ?? []) as string[];
-
-  // Every wiki URL is id-based: `/app/wiki/<id>` (Google-Docs style). The id is
-  // stable, so the URL survives rename/move. A legacy `/app/wiki/<path>` URL
-  // still resolves as input and is redirected to its id URL below.
-  const first = rawSlugParts[0];
-  // An id URL is always a single segment (`/app/wiki/<id>`); a multi-segment
-  // slug is a legacy path even if its first segment looks like an id.
-  const idMode = rawSlugParts.length === 1 && !!first && isDocId(first);
-  const commentId = searchParams?.get("comment") ?? null;
-
-  // id mode: resolve the id to its current path / kind / deleted state.
-  const { data: resolved, error: resolveErr } = useSWR(
-    idMode ? `/wiki/id/${first}` : null,
-    () => resolveDocId(first as string),
-    { revalidateOnFocus: false },
-  );
-
-  // Next.js may hand back percent-encoded segments (e.g. "local%20testing").
-  // Decode so labels and downstream API paths use literal characters.
-  const decodedParts = rawSlugParts.map((s) => {
-    try {
-      return decodeURIComponent(s);
-    } catch {
-      return s;
-    }
-  });
-  const pathModePath = decodedParts.join("/");
-  // A single 16-hex segment is treated as an id. When it doesn't resolve (404)
-  // it's almost always an unknown/expired id — but it *could* be a legacy path
-  // literally named like an id, so we resolve that path: a live id means it's a
-  // real doc (redirect to its canonical id URL below); no id means the id is
-  // simply unknown → the "unavailable" card. This is why an unknown id must NOT
-  // fall through to the folder explorer (which would render an empty folder).
-  const idResolve404 =
-    idMode && (resolveErr as ApiError | undefined)?.status === 404;
-  const {
-    data: idFallback,
-    error: idFallbackError,
-    isLoading: idFallbackLoading,
-  } = useSWR(
-    idResolve404 && pathModePath ? ["id-fallback", pathModePath] : null,
-    () => resolveIds([pathModePath]).then((m) => m[pathModePath] ?? null),
-    { revalidateOnFocus: false },
-  );
-  // The doc/folder path we're viewing: from the resolved id, or (legacy path
-  // URL) straight from the URL. An unresolved id contributes no path.
-  const effectivePath = idMode ? (resolved?.path ?? null) : pathModePath;
-  const isFile = !!effectivePath && effectivePath.endsWith(".md");
-
-  // Legacy path URL (`/app/wiki/<path>`): every real page/folder has a stable
-  // id, so we resolve the path → redirect to its canonical id URL. A path that
-  // resolves to *no* id doesn't exist → the render shows "unavailable" rather
-  // than an empty folder explorer. The wiki root and the new-doc flow have no
-  // path/id and are handled directly.
-  const pathModeActive = !idMode && !!pathModePath && !isNewMode;
-  const {
-    data: pathId,
-    error: pathIdError,
-    isLoading: pathIdLoading,
-  } = useSWR(
-    pathModeActive ? ["wiki-path-id", pathModePath] : null,
-    () => resolveIds([pathModePath]).then((m) => m[pathModePath] ?? null),
-    { revalidateOnFocus: false },
-  );
-  useEffect(() => {
-    if (!pathModeActive || !pathId) return;
-    const suffix = commentId ? `?comment=${encodeURIComponent(commentId)}` : "";
-    router.replace(wikiHref(pathId) + suffix);
-  }, [pathModeActive, pathId, commentId, router]);
-
-  // A 16-hex id that didn't resolve but *is* a real doc's path (a page/folder
-  // literally named like an id) → send it to its canonical id URL, preserving
-  // a comment deep-link like the path redirect does.
-  useEffect(() => {
-    if (!idResolve404 || !idFallback) return;
-    const suffix = commentId ? `?comment=${encodeURIComponent(commentId)}` : "";
-    router.replace(wikiHref(idFallback) + suffix);
-  }, [idResolve404, idFallback, commentId, router]);
-
-  // Remember the most recent wiki path (for the "Last viewed" landing) and
-  // feed the sidebar Recents when an actual doc is opened.
-  useEffect(() => {
-    if (effectivePath === null) return;
-    rememberWikiPath("/app/wiki" + (effectivePath ? "/" + effectivePath : ""));
-    if (isFile) void recordRecentDoc(effectivePath);
-  }, [effectivePath, isFile]);
-
-  // Tab title tracks the open doc (or folder); falls back to the app name.
-  useEffect(() => {
-    if (effectivePath === null) return;
-    const last = effectivePath.split("/").pop() ?? "";
-    const label = isFile ? pageTitle(effectivePath) : last;
-    document.title = label || "agent-wiki";
-    return () => {
-      document.title = "agent-wiki";
-    };
-  }, [effectivePath, isFile]);
-
-  if (loading || !user)
-    return (
-      <main className={isMobile ? "p-4" : "p-8"}>
-        <LoadingSpinner center />
-      </main>
-    );
-
-  if (idMode) {
-    if (idResolve404) {
-      // Unknown id. If it's actually a live doc's path we're mid-redirect (or
-      // still checking) → spinner; otherwise the unavailable card (never the
-      // empty folder explorer). If the fallback lookup itself failed (after
-      // SWR's retries) we can't confirm the path is missing, so show the softer
-      // generic copy rather than the definitive 404 "broken link" claim.
-      if (idFallbackLoading || idFallback)
-        return (
-          <main className={isMobile ? "p-4" : "p-8"}>
-            <LoadingSpinner center />
-          </main>
-        );
-      return <WikiUnknownLink status={idFallbackError ? undefined : 404} />;
-    }
-    // Deleted page → tombstone panel (who/when + Restore); other resolve
-    // failure → the generic unavailable card.
-    if (resolved?.deleted_at) return <WikiTombstone path={resolved.path} />;
-    if (resolveErr)
-      return <WikiUnknownLink status={(resolveErr as ApiError)?.status} />;
-    if (!resolved)
-      return (
-        <main className={isMobile ? "p-4" : "p-8"}>
-          <LoadingSpinner center />
-        </main>
-      );
-  }
-
-  if (pathModeActive) {
-    // Real page/folder → has an id → redirecting to its canonical id URL
-    // (spinner). Still resolving → spinner. Resolved to no id → the path
-    // doesn't exist → unavailable. On a transient resolve failure, fall through
-    // to render-by-path below rather than falsely claim unavailable.
-    if (pathIdLoading || pathId)
-      return (
-        <main className={isMobile ? "p-4" : "p-8"}>
-          <LoadingSpinner center />
-        </main>
-      );
-    if (!pathIdError && pathId === null)
-      return <WikiUnknownLink status={404} />;
-  }
-
-  if (isFile) return <FileView path={effectivePath as string} />;
-  if (isNewMode) return <NewDocView dir={pathModePath} />;
-  // Wiki root with no doc open → the "Welcome to Onyx Wiki" landing.
-  // Sub-folders still render the directory explorer.
-  if (!effectivePath) return <WikiHome />;
-  return <Explorer dir={effectivePath} />;
-}
-
-/** A wiki URL that no longer points at a live doc — an unknown/expired id, or
- * a page that was deleted but is no longer in Trash (purged). A deleted page
- * still in Trash renders {@link WikiTombstone} instead. */
-function WikiUnknownLink({ status }: { status?: number }) {
-  const router = useRouter();
-  return (
-    <main className="flex h-full items-center justify-center p-8 pb-[16vh]">
-      <div className="flex flex-col items-center gap-4">
-        <EmptyMessageCard
-          sizePreset="main-ui"
-          icon={SvgAlertTriangle}
-          title="This page isn’t available"
-          description={
-            status === 404
-              ? "The link may be broken, or the page was removed."
-              : "The page it pointed to may have been moved or removed."
-          }
-        />
-        <Button onClick={() => router.push("/app/wiki")}>
-          Go to wiki home
-        </Button>
-      </div>
-    </main>
-  );
-}
-
-/** Tombstone for a deleted page/folder reached via its id URL: shows who/when
- * and offers Restore. Falls back to {@link WikiUnknownLink} when the item is no
- * longer in Trash (purged, or deleted before Trash shipped). */
-function WikiTombstone({ path }: { path: string }) {
-  const router = useRouter();
-  const [busy, setBusy] = useState(false);
-  const [err, setErr] = useState<string | null>(null);
-  const {
-    data: entry,
-    error,
-    isLoading,
-  } = useSWR(
-    `/wiki/deleted?path=${encodeURIComponent(path)}`,
-    () => getDeletedTombstone(path),
-    { revalidateOnFocus: false },
-  );
-
-  if (isLoading)
-    return (
-      <main className="p-8">
-        <LoadingSpinner center />
-      </main>
-    );
-  if (error || !entry)
-    return <WikiUnknownLink status={(error as ApiError)?.status} />;
-
-  const label = pageTitle(entry.path);
-  const restore = async () => {
-    setBusy(true);
-    setErr(null);
-    try {
-      const res = await restoreTrashed(entry.trash_id);
-      // Land on the restored page; the route resolves the path → id URL.
-      // wikiPath encodes each segment (paths may contain spaces, #, %, …).
-      router.push(wikiPath(res.path));
-    } catch (e) {
-      setErr(
-        e instanceof ApiError && e.status === 409
-          ? `A page already exists at ${entry.path}.`
-          : e instanceof Error
-            ? e.message
-            : "Restore failed.",
-      );
-      setBusy(false);
-    }
-  };
-
-  return (
-    <main className="flex h-full items-center justify-center p-8 pb-[16vh]">
-      <div className="flex flex-col items-center gap-4">
-        <EmptyMessageCard
-          sizePreset="main-ui"
-          icon={SvgTrash}
-          title={`“${label}” was deleted`}
-          description={`Deleted by ${entry.trashed_by}${
-            entry.trashed_at ? ` · ${formatRelative(entry.trashed_at)}` : ""
-          }. Restore it to bring it back to ${entry.path}.`}
-        />
-        {err && <MessageCard variant="error" title={err} />}
-        <div className="flex items-center gap-2">
-          <Button
-            prominence="secondary"
-            onClick={() => router.push("/app/wiki")}
-          >
-            Go to wiki home
-          </Button>
-          {entry.can_restore && (
-            <Button onClick={() => void restore()} disabled={busy}>
-              {busy ? "Restoring…" : "Restore"}
-            </Button>
-          )}
-        </div>
-      </div>
-    </main>
-  );
-}
-
-function Explorer({ dir }: { dir: string }) {
-  const router = useRouter();
-  const isMobile = useIsMobile();
-  const host = useHeaderActionsHost();
-  const {
-    data,
-    error: listError,
-    mutate: mutatePaths,
-  } = useSWR<ListResponse>("/wiki");
-  const entries = data?.entries ?? [];
-  const [mutationError, setMutationError] = useState<string | null>(null);
-  const confirmDialog = useConfirm();
-  const error =
-    mutationError ?? (listError instanceof Error ? listError.message : null);
-  const setError = setMutationError;
-  // Force the cache to revalidate from the server after writes (create /
-  // delete / move). Refreshes this listing *and* every wiki cache — including
-  // the open doc's id→path resolve and content — so a rename/move is reflected
-  // on screen instead of leaving the view on the old path.
-  const refresh = useCallback(() => {
-    void mutatePaths();
-    void revalidateWiki();
-  }, [mutatePaths]);
-  const [busyPath, setBusyPath] = useState<string | null>(null);
-  // Folders still use an inline filename form; new docs route to
-  // NewDocView where filename + template + body are chosen together.
-  const [creating, setCreating] = useState<"folder" | null>(null);
-  const [newName, setNewName] = useState("");
-  const [createBusy, setCreateBusy] = useState(false);
-  const [triggerModalOpen, setTriggerModalOpen] = useState(false);
-  const [triggerStatus, setTriggerStatus] = useState<string | null>(null);
-  const [sort, setSort] = useState<"name-asc" | "name-desc" | "recent">(
-    "name-asc",
-  );
-  const [renaming, setRenaming] = useState<string | null>(null);
-  const [dragSource, setDragSource] = useState<string | null>(null);
-  const [dropTarget, setDropTarget] = useState<string | null>(null);
-  const [sharePath, setSharePath] = useState<string | null>(null);
-  const [policyOpen, setPolicyOpen] = useState(false);
-
-  const { subdirs, files } = useMemo(() => {
-    const prefix = dir ? dir + "/" : "";
-    // Folder mtime = max of descendant entries' timestamps.
-    const dirMtime = new Map<string, string>();
-    const fileList: { name: string; updated_at: string }[] = [];
-    for (const e of entries) {
-      if (!e.path.startsWith(prefix)) continue;
-      const rest = e.path.slice(prefix.length);
-      if (!rest) continue;
-      const slash = rest.indexOf("/");
-      if (slash === -1) {
-        if (rest.endsWith(".md"))
-          fileList.push({ name: rest, updated_at: e.updated_at });
-      } else {
-        const name = rest.slice(0, slash);
-        const cur = dirMtime.get(name);
-        if (!cur || (e.updated_at && e.updated_at > cur)) {
-          dirMtime.set(name, e.updated_at);
-        }
-      }
-    }
-    const dirList = [...dirMtime.entries()].map(([name, updated_at]) => ({
-      name,
-      updated_at,
-    }));
-    const byName =
-      (asc: boolean) => (a: { name: string }, b: { name: string }) =>
-        asc ? a.name.localeCompare(b.name) : b.name.localeCompare(a.name);
-    // Newest first; empty timestamps sink to the bottom.
-    const byRecent = (a: { updated_at: string }, b: { updated_at: string }) => {
-      if (!a.updated_at && !b.updated_at) return 0;
-      if (!a.updated_at) return 1;
-      if (!b.updated_at) return -1;
-      return b.updated_at.localeCompare(a.updated_at);
-    };
-    const cmp = sort === "recent" ? byRecent : byName(sort === "name-asc");
-    return {
-      subdirs: dirList.sort(cmp),
-      files: fileList.sort(cmp),
-    };
-  }, [entries, dir, sort]);
-
-  async function onCreate(e: FormEvent) {
-    e.preventDefault();
-    const raw = newName.trim();
-    if (!raw) return;
-    // Only folders go through this inline form; new docs are handled by
-    // NewDocView, which the +New document button routes to directly.
-    if (creating !== "folder") return;
-    setCreateBusy(true);
-    setError(null);
-    try {
-      const folderName = raw.replace(/\/+$/, "");
-      const fullPath = (dir ? dir + "/" : "") + folderName;
-      await apiFetch("/wiki/folder", {
-        method: "POST",
-        body: JSON.stringify({ path: fullPath }),
-      });
-      setNewName("");
-      setCreating(null);
-      refresh();
-      router.push(`/app/wiki/${fullPath}`);
-    } catch (e) {
-      setError(e instanceof Error ? e.message : "create failed");
-    } finally {
-      setCreateBusy(false);
-    }
-  }
-
-  async function onDelete(rel: string) {
-    if (
-      !(await confirmDialog({
-        title: `Delete ${rel}?`,
-        body: "It will be moved to Trash, where you can restore it.",
-        confirmLabel: "Delete",
-      }))
-    )
-      return;
-    setBusyPath(rel);
-    setError(null);
-    try {
-      await apiFetch(`/wiki/file?path=${encodeURIComponent(rel)}`, {
-        method: "DELETE",
-      });
-      refresh();
-    } catch (e) {
-      setError(e instanceof Error ? e.message : "delete failed");
-    } finally {
-      setBusyPath(null);
-    }
-  }
-
-  async function onMove(srcRel: string, destDir: string) {
-    const base = srcRel.split("/").pop() ?? srcRel;
-    const newRel = destDir ? `${destDir}/${base}` : base;
-    if (newRel === srcRel) return;
-    // Block dropping a folder onto itself or any of its descendants.
-    if (destDir === srcRel || destDir.startsWith(srcRel + "/")) {
-      setError("Cannot move a folder into itself.");
-      return;
-    }
-    setBusyPath(srcRel);
-    setError(null);
-    try {
-      await apiFetch("/wiki/move", {
-        method: "POST",
-        body: JSON.stringify({ old_path: srcRel, new_path: newRel }),
-      });
-      refresh();
-    } catch (e) {
-      setError(e instanceof Error ? e.message : "move failed");
-    } finally {
-      setBusyPath(null);
-    }
-  }
-
-  async function onRenameSubmit(rel: string, rawName: string) {
-    const trimmed = rawName.trim().replace(/^\/+|\/+$/g, "");
-    if (!trimmed || trimmed.includes("/")) {
-      setError("Name cannot be empty or contain '/'.");
-      return;
-    }
-    const segs = rel.split("/");
-    const parent = segs.slice(0, -1).join("/");
-    const isFile = rel.endsWith(".md");
-    const finalName =
-      isFile && !trimmed.endsWith(".md") ? trimmed + ".md" : trimmed;
-    const newRel = parent ? `${parent}/${finalName}` : finalName;
-    if (newRel === rel) {
-      setRenaming(null);
-      return;
-    }
-    setBusyPath(rel);
-    setError(null);
-    try {
-      await apiFetch("/wiki/move", {
-        method: "POST",
-        body: JSON.stringify({ old_path: rel, new_path: newRel }),
-      });
-      setRenaming(null);
-      refresh();
-    } catch (e) {
-      setError(e instanceof Error ? e.message : "rename failed");
-    } finally {
-      setBusyPath(null);
-    }
-  }
-
-  // Folder-level actions portal into the single pinned header as icon buttons;
-  // the breadcrumb there is the shared one (WikiHeader), so this view renders
-  // no header of its own.
-  const headerActions = (
-    <>
-      <Button
-        icon={SvgWorkflow}
-        prominence="tertiary"
-        tooltip="Trigger"
-        onClick={() => setTriggerModalOpen(true)}
-      />
-      <Button
-        icon={SvgShield}
-        prominence="tertiary"
-        tooltip="Update Policy"
-        onClick={() => setPolicyOpen(true)}
-      />
-      <Button
-        icon={SvgFolderPlus}
-        prominence="tertiary"
-        tooltip="New folder"
-        onClick={() => {
-          setNewName("");
-          setCreating((v) => (v === "folder" ? null : "folder"));
-        }}
-      />
-      <Button
-        icon={SvgPlus}
-        variant="action"
-        tooltip="New document"
-        onClick={() => router.push(`/app/wiki/${dir}?new=1`)}
-      />
-    </>
-  );
-
-  return (
-    <main
-      className={`h-full overflow-y-auto ${isMobile ? "px-3 py-4" : "px-8 py-6"}`}
-    >
-      {host?.el && createPortal(headerActions, host.el)}
-
-      <TriggerPanel
-        open={triggerModalOpen}
-        initial={{ scope_path: dir || "/" }}
-        lockScope
-        onClose={() => setTriggerModalOpen(false)}
-        onSaved={(t) => setTriggerStatus(`Created trigger for ${t.scope_path}`)}
-      />
-
-      {triggerStatus && (
-        <div className="mb-3 text-xs text-(--text-04)">{triggerStatus}</div>
-      )}
-
-      {creating && (
-        <form
-          onSubmit={onCreate}
-          className="mb-4 flex gap-2 rounded-(--border-radius-08) border border-(--border-01) bg-(--background-tint-01) p-3"
-        >
-          <input
-            autoFocus
-            value={newName}
-            onChange={(e) => setNewName(e.target.value)}
-            placeholder="folder-name (or subdir/folder-name)"
-            disabled={createBusy}
-            className="flex-1 rounded-(--border-radius-04) border border-(--border-01) p-2 text-sm"
-          />
-          <Button
-            type="submit"
-            variant="action"
-            disabled={createBusy || !newName.trim()}
-          >
-            Create folder
-          </Button>
-          <Button
-            type="button"
-            onClick={() => {
-              setCreating(null);
-              setNewName("");
-            }}
-          >
-            Cancel
-          </Button>
-        </form>
-      )}
-
-      {error && (
-        <div className="mb-3 rounded-(--border-radius-04) bg-(--status-error-01) p-[10px] text-[13px] text-(--status-text-error-05)">
-          {error}
-        </div>
-      )}
-
-      {subdirs.length === 0 && files.length === 0 && !error && (
-        <p className="text-sm text-(--text-03)">
-          This folder is empty. Create a document to get started.
-        </p>
-      )}
-
-      {(subdirs.length > 0 || files.length > 0) && (
-        <SortBar value={sort} onChange={setSort} />
-      )}
-
-      <ul className="m-0 list-none p-0">
-        {(() => {
-          const dirEntries = subdirs.map((d) => ({ ...d, isFile: false }));
-          const fileEntries = files.map((f) => ({ ...f, isFile: true }));
-          // Folders always above docs; ordering within each group is set by `sort`.
-          const ordered = [...dirEntries, ...fileEntries];
-          return ordered.map(({ name, updated_at, isFile }) => {
-            const childPath = (dir ? dir + "/" : "") + name;
-            return (
-              <Row
-                key={(isFile ? "f:" : "d:") + name}
-                icon={
-                  isFile ? (
-                    <SvgDocFile size={20} aria-hidden />
-                  ) : (
-                    <SvgFolder size={20} aria-hidden />
-                  )
-                }
-                label={name}
-                updatedAt={updated_at}
-                href={`/app/wiki/${childPath}`}
-                path={childPath}
-                isFile={isFile}
-                busy={busyPath === childPath}
-                onDelete={() => onDelete(childPath)}
-                onShare={() => setSharePath(childPath)}
-                renaming={renaming === childPath}
-                onStartRename={() => setRenaming(childPath)}
-                onCancelRename={() => setRenaming(null)}
-                onSubmitRename={(v) => onRenameSubmit(childPath, v)}
-                onDragStart={() => setDragSource(childPath)}
-                onDragEnd={() => {
-                  setDragSource(null);
-                  setDropTarget(null);
-                }}
-                dropActive={!isFile && dropTarget === childPath}
-                onFolderDragOver={
-                  isFile
-                    ? undefined
-                    : () => {
-                        if (dragSource && dragSource !== childPath) {
-                          setDropTarget(childPath);
-                        }
-                      }
-                }
-                onFolderDragLeave={
-                  isFile
-                    ? undefined
-                    : () =>
-                        setDropTarget((cur) => (cur === childPath ? null : cur))
-                }
-                onFolderDrop={
-                  isFile
-                    ? undefined
-                    : () => {
-                        if (dragSource && dragSource !== childPath) {
-                          onMove(dragSource, childPath);
-                        }
-                        setDragSource(null);
-                        setDropTarget(null);
-                      }
-                }
-              />
-            );
-          });
-        })()}
-      </ul>
-      {sharePath && (
-        <ShareDialog path={sharePath} open onClose={() => setSharePath(null)} />
-      )}
-      {policyOpen && (
-        // Folder policy applies to every page under this folder. Shown as a
-        // right-side drawer (the folder listing isn't a flex row like the page
-        // view), reusing the same panel — ``dir`` is the folder path ("" = root).
-        <>
-          <div
-            onClick={() => setPolicyOpen(false)}
-            aria-hidden
-            className="fixed inset-0 z-[60] bg-(--mask-03)"
-          />
-          <div className="fixed top-0 right-0 bottom-0 z-[70] flex w-[min(400px,100vw)] shadow-(--shadow-panel)">
-            <UpdatePolicyPanel
-              path={dir}
-              onClose={() => setPolicyOpen(false)}
-              fullHeight
-            />
-          </div>
-        </>
-      )}
-    </main>
-  );
-}
-
-function NewDocView({ dir }: { dir: string }) {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const isMobile = useIsMobile();
-  const host = useHeaderActionsHost();
-  const { setDrafting, requestExpand, registerDraftBridge } = useDrafting();
-  // "Start writing with AI" hands a generated draft (+ the prompt) here via
-  // sessionStorage, paired with ?ai=1. Read it synchronously so the editor and
-  // the drafting chat seed on the first render (no blank→ai re-init flash).
-  const isAiSeed = searchParams?.get("ai") === "1";
-  const [aiSeed] = useState<{
-    title?: string;
-    body?: string;
-    prompt?: string;
-  } | null>(() => {
-    if (!isAiSeed || typeof window === "undefined") return null;
-    try {
-      const raw = sessionStorage.getItem(AI_DRAFT_KEY);
-      return raw ? JSON.parse(raw) : null;
-    } catch {
-      return null;
-    }
-  });
-  const [filename, setFilename] = useState(aiSeed?.title ?? "");
-  const [draft, setDraft] = useState(aiSeed?.body ?? "");
-  const [templates, setTemplates] = useState<DocumentTemplateSummary[] | null>(
-    null,
-  );
-  const [appliedTemplateId, setAppliedTemplateId] = useState<string | null>(
-    null,
-  );
-  const [appliedTemplateBody, setAppliedTemplateBody] = useState<string | null>(
-    null,
-  );
-  const [applyingTemplateId, setApplyingTemplateId] = useState<string | null>(
-    null,
-  );
-  const [saving, setSaving] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  // Destination folder for the new doc — defaults to the route's folder but is
-  // user-selectable (the AI flow lands at root, so the picker is how you place it).
-  const [destDir, setDestDir] = useState(dir);
-  const { data: tree } = useSWR<ListResponse>("/wiki");
-  const folders = useMemo(() => collectFolders(tree?.entries ?? []), [tree]);
-
-  // Drop the stash once consumed so it doesn't re-apply on remount / back-nav.
-  useEffect(() => {
-    if (aiSeed && typeof window !== "undefined") {
-      try {
-        sessionStorage.removeItem(AI_DRAFT_KEY);
-      } catch {
-        // ignore
-      }
-    }
-  }, [aiSeed]);
-
-  // Bridge the editor to the drafting chat so it can live-edit this unsaved
-  // draft. Keep a ref of the latest body so the chat reads current content.
-  const draftRef = useRef(draft);
-  useEffect(() => {
-    draftRef.current = draft;
-  }, [draft]);
-  useEffect(() => {
-    registerDraftBridge({ get: () => draftRef.current, set: setDraft });
-    return () => registerDraftBridge(null);
-  }, [registerDraftBridge]);
-
-  // Pop the chat widget open once on mount — the assistant can help while the
-  // user drafts. For an AI-seeded draft the chat seeds with the user's prompt.
-  useEffect(() => {
-    requestExpand();
-  }, [requestExpand]);
-
-  useEffect(() => {
-    let cancelled = false;
-    listTemplateSummaries()
-      .then((rows) => {
-        if (!cancelled) setTemplates(rows);
-      })
-      .catch(() => {
-        if (!cancelled) setTemplates([]);
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, []);
-
-  // Sync drafting context with the current pick — including the
-  // initial "no template chosen yet" state, which maps to ``blank``
-  // drafting so the chat kicks off the moment +New routes into this
-  // view. Picking a template later swaps ``desiredKey`` from "blank"
-  // to "tpl:<id>" and the chat widget re-inits a fresh session for it.
-  useEffect(() => {
-    if (appliedTemplateId) {
-      if (!templates) return; // wait until we can resolve the name
-      const t = templates.find((x) => x.id === appliedTemplateId);
-      setDrafting({
-        kind: "template",
-        path: null,
-        templateId: appliedTemplateId,
-        templateName: t?.name ?? null,
-      });
-    } else {
-      setDrafting({ kind: "blank", path: null, prompt: aiSeed?.prompt });
-    }
-  }, [appliedTemplateId, templates, setDrafting, aiSeed]);
-  // Clear drafting on unmount (cancel, sidebar nav, …) — the chat widget
-  // tears its drafting session down synchronously on null, so the collapse
-  // happens in the same paint as the page change. The one exception is
-  // Create: it navigates to the doc it just made and FileView re-syncs
-  // drafting from the server-side draft row, so passing through null there
-  // would collapse the chat only to re-init it a moment later.
-  const createHandoffRef = useRef(false);
-  useEffect(() => {
-    return () => {
-      if (!createHandoffRef.current) setDrafting(null);
-    };
-  }, [setDrafting]);
-
-  const trimmedFilename = filename.trim().replace(/^\/+|\/+$/g, "");
-  const filenameNoExt = trimmedFilename.replace(/\.md$/i, "");
-  const filenameValid = !!filenameNoExt && !filenameNoExt.includes("/");
-  // Block Create while a template's body is still loading, so it can't be
-  // saved before the template (and its policy) is applied.
-  const canCreate = filenameValid && !saving && applyingTemplateId === null;
-
-  async function onPickTemplate(template: DocumentTemplateSummary) {
-    setApplyingTemplateId(template.id);
-    // Set the applied id synchronously — the create request seeds the page's
-    // policy from it, so it must be set before the (async) body load, not after.
-    setAppliedTemplateId(template.id);
-    setError(null);
-    try {
-      const full = await getTemplate(template.id);
-      setDraft(full.body);
-      setAppliedTemplateBody(full.body);
-    } catch (e) {
-      setError(e instanceof Error ? e.message : "failed to apply template");
-    } finally {
-      setApplyingTemplateId(null);
-    }
-  }
-
-  function onPickBlank() {
-    setDraft("");
-    setAppliedTemplateBody(null);
-    setAppliedTemplateId(null);
-    // Kick the chat widget into blank-drafting mode so it spins up a
-    // hidden session with the generic "what would you like to work on"
-    // prime, the same way a template pick spins up a template-aware
-    // session.
-    setDrafting({ kind: "blank", path: null });
-  }
-
-  // Deep-link from the home page: ``?template=<id>`` pre-applies that
-  // template once the summaries load, so the landing's template cards
-  // drop the user straight into a seeded draft.
-  const templateParam = searchParams?.get("template") ?? null;
-  const autoTemplateRef = useRef(false);
-  useEffect(() => {
-    if (autoTemplateRef.current) return;
-    if (!templateParam || !templates) return;
-    const match = templates.find((t) => t.id === templateParam);
-    if (!match) return;
-    autoTemplateRef.current = true;
-    void onPickTemplate(match);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [templateParam, templates]);
-
-  async function onCreate() {
-    if (!filenameValid) {
-      setError("Filename cannot be empty or contain '/'.");
-      return;
-    }
-    setSaving(true);
-    setError(null);
-    try {
-      const name = filenameNoExt + ".md";
-      const fullPath = (destDir ? destDir + "/" : "") + name;
-      const created = await apiFetch<{ id?: string | null }>("/wiki/file", {
-        method: "PUT",
-        body: JSON.stringify({
-          path: fullPath,
-          body: draft,
-          // Seeds the new page's update policy from the template; the draft
-          // recorded below lands after the commit, too late for that.
-          ...(appliedTemplateId ? { template_id: appliedTemplateId } : {}),
-        }),
-      });
-      // If a template was applied, record the draft row so the chat
-      // banner + template system prompt persist on the saved doc.
-      if (appliedTemplateId) {
-        await setDraftTemplate(fullPath, appliedTemplateId);
-      }
-      // Revalidate every wiki cache so the persistent Directory sidebar (and
-      // any open folder listing) shows the new page without a full reload.
-      void revalidateWiki();
-      // Hand-off: keep the drafting state (and the chat's drafting
-      // session) alive across the navigation — see the unmount cleanup.
-      createHandoffRef.current = true;
-      // Land on the new page's id URL directly (the create response carries
-      // the minted id), so it's a clean /app/wiki/<id> like every other page.
-      // Keep ?new=1 — FileView reads it to auto-open the assistant on a
-      // freshly-created doc.
-      const base = created.id ? wikiHref(created.id) : `/app/wiki/${fullPath}`;
-      router.push(`${base}?new=1`);
-    } catch (e) {
-      setError(e instanceof Error ? e.message : "create failed");
-      setSaving(false);
-    }
-  }
-
-  const isBlank = draft.trim() === "";
-  const matchesApplied =
-    appliedTemplateBody !== null && draft === appliedTemplateBody;
-  const showGallery =
-    (isBlank || matchesApplied) && templates !== null && templates.length > 0;
-  const parentSlug = destDir;
-
-  // Cancel / Create portal into the single pinned header; the breadcrumb there
-  // shows the destination folder, so this view renders no header of its own.
-  const headerActions = (
-    <>
-      <Button onClick={() => router.push(`/app/wiki/${dir}`)} disabled={saving}>
-        Cancel
-      </Button>
-      <Button
-        variant="action"
-        onClick={() => void onCreate()}
-        disabled={!canCreate}
-        tooltip={
-          !filenameValid && !saving ? "Give the file a name first." : undefined
-        }
-      >
-        {saving ? "Creating…" : "Create"}
-      </Button>
-    </>
-  );
-
-  return (
-    <main
-      className={`box-border flex h-full flex-col gap-3 ${isMobile ? "px-3 py-4" : "px-8 py-6"}`}
-    >
-      {host?.el && createPortal(headerActions, host.el)}
-
-      <div className="flex shrink-0 items-center gap-2">
-        <span className="text-[13px] text-(--text-04)">Folder</span>
-        <DestinationSelect
-          value={destDir}
-          folders={folders}
-          onChange={setDestDir}
-          disabled={saving}
-        />
-      </div>
-
-      <FilenameRow
-        parent={parentSlug}
-        value={filename}
-        onChange={setFilename}
-        disabled={saving}
-        autoFocus
-        placeholder="filename for the new doc"
-      />
-
-      {error && (
-        <div className="rounded-(--border-radius-04) bg-(--status-error-01) p-[10px] text-[13px] text-(--status-text-error-05)">
-          {error}
-        </div>
-      )}
-
-      {showGallery && (
-        <TemplateGallery
-          templates={templates!}
-          activeId={matchesApplied ? appliedTemplateId : null}
-          applyingId={applyingTemplateId}
-          blankActive={isBlank && appliedTemplateId === null}
-          onPick={(t) => void onPickTemplate(t)}
-          onBlank={onPickBlank}
-        />
-      )}
-
-      <textarea
-        value={draft}
-        onChange={(e) => setDraft(e.target.value)}
-        spellCheck={false}
-        placeholder={
-          templates && templates.length > 0
-            ? "Start typing, or pick a template above…"
-            : "Start typing your new document…"
-        }
-        className="box-border min-h-0 w-full flex-1 resize-none rounded-(--border-radius-08) border border-(--border-01) p-4 font-mono text-sm leading-[1.6] outline-none"
-      />
-    </main>
-  );
-}
-
-function TemplateGallery({
-  templates,
-  activeId,
-  applyingId,
-  blankActive,
-  onPick,
-  onBlank,
-}: {
-  templates: DocumentTemplateSummary[];
-  activeId: string | null;
-  applyingId: string | null;
-  blankActive: boolean;
-  onPick: (t: DocumentTemplateSummary) => void;
-  onBlank: () => void;
-}) {
-  // Always a single-row strip — the picker never wraps to a second
-  // line. On wide screens the user scrolls / clicks chevrons through
-  // the row; on narrow screens the same layout becomes a swipe strip.
-  return (
-    <div className="flex flex-col gap-[10px] rounded-(--border-radius-08) border border-(--border-01) bg-(--background-tint-01) p-[14px]">
-      <div className="flex items-baseline gap-2">
-        <span className="text-[13px] font-semibold text-(--text-05)">
-          Start from a template
-        </span>
-        <span className="text-xs text-(--text-03)">
-          Scroll or use the arrows to browse — tap to apply.
-        </span>
-      </div>
-      <TemplateStrip
-        templates={templates}
-        activeId={activeId}
-        applyingId={applyingId}
-        blankActive={blankActive}
-        onPick={onPick}
-        onBlank={onBlank}
-      />
-    </div>
-  );
-}
-
-function TemplateStrip({
-  templates,
-  activeId,
-  applyingId,
-  blankActive,
-  onPick,
-  onBlank,
-}: {
-  templates: DocumentTemplateSummary[];
-  activeId: string | null;
-  applyingId: string | null;
-  blankActive: boolean;
-  onPick: (t: DocumentTemplateSummary) => void;
-  onBlank: () => void;
-}) {
-  const scrollerRef = useRef<HTMLDivElement | null>(null);
-  const [edges, setEdges] = useState<{ left: boolean; right: boolean }>({
-    left: false,
-    right: false,
-  });
-
-  const recomputeEdges = useCallback(() => {
-    const el = scrollerRef.current;
-    if (!el) return;
-    const atStart = el.scrollLeft <= 1;
-    const atEnd = el.scrollLeft + el.clientWidth >= el.scrollWidth - 1;
-    setEdges({ left: !atStart, right: !atEnd });
-  }, []);
-
-  useEffect(() => {
-    recomputeEdges();
-    const el = scrollerRef.current;
-    if (!el) return;
-    el.addEventListener("scroll", recomputeEdges, { passive: true });
-    window.addEventListener("resize", recomputeEdges);
-    return () => {
-      el.removeEventListener("scroll", recomputeEdges);
-      window.removeEventListener("resize", recomputeEdges);
-    };
-  }, [recomputeEdges, templates.length]);
-
-  const scrollBy = useCallback((dx: number) => {
-    const el = scrollerRef.current;
-    if (!el) return;
-    el.scrollBy({ left: dx, behavior: "smooth" });
-  }, []);
-
-  // One full card width per click feels right — the user advances by
-  // a card rather than by a viewport, so they never lose their place.
-  const CARD_WIDTH = 200;
-  const STEP = CARD_WIDTH + 8; // card width + gap
-
-  // The "Blank" template is the empty-start entry point — route the blank card
-  // through it (so its auto-update policy applies) when present, and drop it
-  // from the list so it isn't shown twice. Falls back to a plain blank page
-  // when no Blank template is seeded.
-  const blankTemplate = templates.find((t) => t.name === "Blank");
-  const rest = templates.filter((t) => t.name !== "Blank");
-  const blankCardActive =
-    blankActive || (blankTemplate != null && activeId === blankTemplate.id);
-
-  return (
-    <div className="relative">
-      <div
-        ref={scrollerRef}
-        className="scroll-x-hidden flex snap-x snap-mandatory gap-2 overflow-x-auto pb-[2px]"
-        style={{ WebkitOverflowScrolling: "touch" }}
-      >
-        <div className="w-[200px] shrink-0 snap-start">
-          <TemplateCard
-            title="Blank document"
-            description="Empty file — just start typing."
-            active={blankCardActive}
-            busy={false}
-            onClick={() => (blankTemplate ? onPick(blankTemplate) : onBlank())}
-          />
-        </div>
-        {rest.map((t) => (
-          <div key={t.id} className="w-[200px] shrink-0 snap-start">
-            <TemplateCard
-              title={t.name}
-              description={t.description}
-              note={
-                t.ingestion_auto_update_disabled ? "Auto-update off" : undefined
-              }
-              active={activeId === t.id}
-              busy={applyingId === t.id}
-              onClick={() => onPick(t)}
-            />
-          </div>
-        ))}
-      </div>
-      {edges.left && (
-        <StripArrow direction="left" onClick={() => scrollBy(-STEP)} />
-      )}
-      {edges.right && (
-        <StripArrow direction="right" onClick={() => scrollBy(STEP)} />
-      )}
-    </div>
-  );
-}
-
-function StripArrow({
-  direction,
-  onClick,
-}: {
-  direction: "left" | "right";
-  onClick: () => void;
-}) {
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      aria-label={direction === "left" ? "Scroll left" : "Scroll right"}
-      className={`absolute top-1/2 -translate-y-1/2 ${direction === "left" ? "left-1" : "right-1"} flex h-7 w-7 cursor-pointer items-center justify-center rounded-full border border-(--border-01) bg-(--background-tint-00) p-0 text-(--text-04) shadow-(--shadow-sm)`}
-    >
-      {direction === "left" ? (
-        <SvgChevronLeft size={14} />
-      ) : (
-        <SvgChevronRight size={14} />
-      )}
-    </button>
-  );
-}
-
-function TemplateCard({
-  title,
-  description,
-  note,
-  active,
-  busy,
-  onClick,
-}: {
-  title: string;
-  description: string | null;
-  note?: string;
-  active: boolean;
-  busy: boolean;
-  onClick: () => void;
-}) {
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      disabled={busy}
-      className={`box-border flex h-full min-h-[64px] w-full flex-col gap-1 rounded-(--border-radius-04) border px-3 py-[10px] text-left text-(--text-05) transition-[background,border-color] duration-[80ms] ease-in-out ${busy ? "cursor-wait opacity-[0.7]" : "cursor-pointer"} ${active ? "border-(--border-01) bg-(--background-tint-03)" : "border-(--border-01) bg-(--background-tint-00)"}`}
-      onMouseEnter={(e) => {
-        if (!active && !busy) {
-          e.currentTarget.style.background = "var(--background-tint-03)";
-          e.currentTarget.style.borderColor = "var(--border-02)";
-        }
-      }}
-      onMouseLeave={(e) => {
-        if (!active && !busy) {
-          e.currentTarget.style.background = "";
-          e.currentTarget.style.borderColor = "";
-        }
-      }}
-    >
-      <div className="text-[13px] font-semibold">{title}</div>
-      {description && (
-        <div className="line-clamp-2 text-xs text-(--text-03)">
-          {description}
-        </div>
-      )}
-      {note && <div className="text-[11px] text-(--text-02)">{note}</div>}
-    </button>
-  );
-}
-
-type SortMode = "name-asc" | "name-desc" | "recent";
-
-function SortBar({
-  value,
-  onChange,
-}: {
-  value: SortMode;
-  onChange: (v: SortMode) => void;
-}) {
-  return (
-    <div className="mb-2 flex items-center gap-2 text-xs text-(--text-03)">
-      <label htmlFor="wiki-sort">Sort:</label>
-      <select
-        id="wiki-sort"
-        value={value}
-        onChange={(e) => onChange(e.target.value as SortMode)}
-        className="rounded-(--border-radius-04) border border-(--border-01) bg-(--background-tint-00) px-2 py-1 text-xs text-(--text-05)"
-      >
-        <option value="name-asc">Name (A → Z)</option>
-        <option value="name-desc">Name (Z → A)</option>
-        <option value="recent">Recently updated</option>
-      </select>
-    </div>
-  );
-}
-
-function Row({
-  icon,
-  label,
-  updatedAt,
-  href,
-  path,
-  isFile,
-  busy,
-  onDelete,
-  onShare,
-  renaming,
-  onStartRename,
-  onCancelRename,
-  onSubmitRename,
-  onDragStart,
-  onDragEnd,
-  dropActive,
-  onFolderDragOver,
-  onFolderDragLeave,
-  onFolderDrop,
-}: {
-  icon: React.ReactNode;
-  label: string;
-  updatedAt: string;
-  href: string;
-  path: string;
-  isFile: boolean;
-  busy: boolean;
-  onDelete: () => void;
-  onShare?: () => void;
-  renaming: boolean;
-  onStartRename: () => void;
-  onCancelRename: () => void;
-  onSubmitRename: (newName: string) => void;
-  onDragStart: () => void;
-  onDragEnd: () => void;
-  dropActive: boolean;
-  onFolderDragOver?: () => void;
-  onFolderDragLeave?: () => void;
-  onFolderDrop?: () => void;
-}) {
-  const router = useRouter();
-  const [hover, setHover] = useState(false);
-  const [draft, setDraft] = useState(label);
-
-  useEffect(() => {
-    if (renaming) setDraft(label);
-  }, [renaming, label]);
-
-  // The whole row acts as the click target *and* the drag source.
-  // Clicks navigate via router.push(href) instead of relying on a
-  // child <Link>, which previously left a dead zone around the icon
-  // and trailing whitespace where the cursor showed "grab" but didn't
-  // navigate. Drags from the action buttons (rename/delete) are
-  // suppressed so a careless drag near the right edge doesn't kick off
-  // a move operation; their clicks stop propagation so they don't
-  // double-fire row navigation.
-  return (
-    <li
-      draggable={!renaming}
-      onClick={(e) => {
-        if (renaming) return;
-        // If the click landed on a button or inside the rename form,
-        // let that element handle it — we don't want delete/rename
-        // taps to also navigate into the doc.
-        const target = e.target as HTMLElement;
-        if (target.closest("button, form, input")) return;
-        router.push(href);
-      }}
-      onDragStart={(e) => {
-        // Cancel drags that start on the action buttons so the user
-        // can mash on rename/delete without yanking the row into a
-        // move state.
-        const target = e.target as HTMLElement;
-        if (target.closest("button, form, input")) {
-          e.preventDefault();
-          return;
-        }
-        e.dataTransfer.effectAllowed = "move";
-        e.dataTransfer.setData("text/plain", path);
-        onDragStart();
-      }}
-      onDragEnd={onDragEnd}
-      onDragOver={
-        onFolderDragOver
-          ? (e) => {
-              e.preventDefault();
-              e.dataTransfer.dropEffect = "move";
-              onFolderDragOver();
-            }
-          : undefined
-      }
-      onDragLeave={onFolderDragLeave}
-      onDrop={
-        onFolderDrop
-          ? (e) => {
-              e.preventDefault();
-              onFolderDrop();
-            }
-          : undefined
-      }
-      onMouseEnter={() => setHover(true)}
-      onMouseLeave={() => setHover(false)}
-      className={`flex items-center border-b border-(--border-01) px-3 py-[10px] ${dropActive ? "bg-(--background-tint-03) outline outline-2 outline-(--background-tint-inverted-00)" : hover ? "bg-(--background-tint-02)" : "bg-transparent"} ${busy ? "opacity-50" : "opacity-100"} ${renaming ? "cursor-default" : "cursor-pointer"}`}
-    >
-      <span className="mr-[10px] flex text-(--text-03)">{icon}</span>
-      {renaming ? (
-        <form
-          onSubmit={(e) => {
-            e.preventDefault();
-            onSubmitRename(draft);
-          }}
-          className="flex flex-1 gap-1.5"
-        >
-          <input
-            autoFocus
-            value={draft}
-            onChange={(e) => setDraft(e.target.value)}
-            onKeyDown={(e) => {
-              if (e.key === "Escape") {
-                e.preventDefault();
-                onCancelRename();
-              }
-            }}
-            disabled={busy}
-            className="flex-1 rounded-(--border-radius-04) border border-(--border-01) px-2 py-1 text-sm"
-          />
-          <Button
-            type="submit"
-            size="sm"
-            variant="action"
-            disabled={busy || !draft.trim()}
-          >
-            Save
-          </Button>
-          <Button
-            type="button"
-            size="sm"
-            onClick={onCancelRename}
-            disabled={busy}
-          >
-            Cancel
-          </Button>
-        </form>
-      ) : (
-        // The label is a plain span — the click target is the parent
-        // <li>. flex: 1 keeps it stretching to fill the space between
-        // icon and action buttons, so any click on the label area
-        // still hits the row's onClick.
-        <span className="flex flex-1 items-center gap-[10px] text-sm text-(--text-05)">
-          {label}
-        </span>
-      )}
-      {!renaming && (
-        <>
-          <span className="mr-2 text-xs whitespace-nowrap text-(--text-02)">
-            {updatedAt ? relativeTime(updatedAt, "short") : "—"}
-          </span>
-          {onShare && (
-            <button
-              onClick={onShare}
-              disabled={busy}
-              title="Share"
-              aria-label={`Share ${label}`}
-              className={`flex items-center border-none bg-transparent p-[6px] ${busy ? "cursor-not-allowed" : "cursor-pointer"} ${hover ? "text-(--text-04)" : "text-transparent"}`}
-            >
-              <SvgShare size={16} />
-            </button>
-          )}
-          <button
-            onClick={onStartRename}
-            disabled={busy}
-            title="Rename"
-            aria-label={`Rename ${label}`}
-            className={`flex items-center border-none bg-transparent p-[6px] ${busy ? "cursor-not-allowed" : "cursor-pointer"} ${hover ? "text-(--text-04)" : "text-transparent"}`}
-          >
-            <SvgEdit size={16} />
-          </button>
-          <button
-            onClick={onDelete}
-            disabled={busy}
-            title="Delete"
-            aria-label={`Delete ${label}`}
-            className={`flex items-center border-none bg-transparent p-[6px] ${busy ? "cursor-not-allowed" : "cursor-pointer"} ${hover ? "text-(--status-text-error-05)" : "text-transparent"}`}
-          >
-            <SvgTrash size={16} />
-          </button>
-        </>
-      )}
-    </li>
-  );
-}
-
-function FileView({ path }: { path: string }) {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const isMobile = useIsMobile();
-  const host = useHeaderActionsHost();
-  const agentsBarHost = useAgentsBarHost();
-  const rightHost = useRightPanelHost();
-  const { isActivitiesOpen, toggleActivities } = useLeftPanel();
-  const { refresh: refreshTriggers } = useTriggers();
-  const { setDrafting, requestExpand } = useDrafting();
-  const { user } = useAuth();
-  const [body, setBody] = useState("");
-  const [editing, setEditing] = useState(false);
-  const [filenameDraft, setFilenameDraft] = useState("");
-  const [saving, setSaving] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [triggerModalOpen, setTriggerModalOpen] = useState(false);
-  const [triggerStatus, setTriggerStatus] = useState<string | null>(null);
-  const [automationsOpen, setAutomationsOpen] = useState(false);
-  const [editingTrigger, setEditingTrigger] = useState<Trigger | null>(null);
-  const [runAgentOpen, setRunAgentOpen] = useState(false);
-  const [shareOpen, setShareOpen] = useState(false);
-  const confirmDialog = useConfirm();
-  // History state. `viewingSha` is null when looking at the working-tree
-  // (latest) version; otherwise it's the sha being viewed and is what we
-  // pass back as `base_sha` on save so the server records a rollback.
-  const [headSha, setHeadSha] = useState<string | null>(null);
-  const [viewingSha, setViewingSha] = useState<string | null>(null);
-  const [historyOpen, setHistoryOpen] = useState(false);
-  const [commits, setCommits] = useState<CommitInfo[] | null>(null);
-  const [diffData, setDiffData] = useState<FileDiffResponse | null>(null);
-  const [historyError, setHistoryError] = useState<string | null>(null);
-  // Comments (render-mode). `commentDraft` is a pending text selection being
-  // composed; `selTool` is the floating "Comment" affordance shown on select.
-  const [commentsOpen, setCommentsOpen] = useState(false);
-  const [policyOpen, setPolicyOpen] = useState(false);
-  const [commentDraft, setCommentDraft] = useState<CommentDraft | null>(null);
-  const [commentThreads, setCommentThreads] = useState<CommentThreadView[]>([]);
-  const [activeCommentId, setActiveCommentId] = useState<string | null>(null);
-  const [selTool, setSelTool] = useState<{
-    x: number;
-    y: number;
-    draft: CommentDraft;
-  } | null>(null);
-  const articleRef = useRef<HTMLElement | null>(null);
-  // Page owns the comment threads (so highlights render even with the panel
-  // closed). Auto-open the panel once per path when a page has comments.
-  const autoOpenedPathRef = useRef<string | null>(null);
-  // A `?comment=<id>` deep-link is focused once per id (reset on path change).
-  const focusedCommentRef = useRef<string | null>(null);
-
-  // The history and comments side panels are mutually exclusive — every
-  // path that opens one closes the other (toolbar toggles, the comments
-  // auto-open, the selection "💬 Comment" tool).
-  const openComments = useCallback(() => {
-    setHistoryOpen(false);
-    setPolicyOpen(false);
-    setAutomationsOpen(false);
-    setTriggerModalOpen(false);
-    setEditingTrigger(null);
-    setCommentsOpen(true);
-  }, []);
-
-  const refreshComments = useCallback(async () => {
-    try {
-      const t = await listComments(path);
-      setCommentThreads(t);
-      // Auto-open once per page only when something needs attention — i.e. an
-      // unresolved thread (open or orphaned, matching the panel's main list).
-      // A page whose comments are all resolved stays closed.
-      const hasUnresolved = t.some(
-        (thread) => thread.root.status !== "resolved",
-      );
-      if (hasUnresolved && autoOpenedPathRef.current !== path) {
-        autoOpenedPathRef.current = path;
-        openComments();
-      }
-    } catch {
-      // comments are non-critical chrome; ignore load failures
-    }
-  }, [path, openComments]);
-
-  useEffect(() => {
-    autoOpenedPathRef.current = null;
-    focusedCommentRef.current = null;
-    setCommentThreads([]);
-    void refreshComments();
-  }, [refreshComments]);
-
-  // MarkdownRenderer is React.memo'd on body — CSS Highlight API ranges painted
-  // over its text nodes stay valid across unrelated re-renders.
-  const renderedBody = <MarkdownRenderer body={body} />;
-
-  // Paint Google-Docs-style highlights over the rendered article for each
-  // (non-orphaned) anchored comment. Cleared in edit/diff mode (no article).
-  //
-  // On a fresh load the comments and the doc body arrive on separate renders,
-  // and react-markdown commits its text nodes a tick later than this effect
-  // runs — so a single synchronous (or even rAF-deferred) paint finds no text
-  // to range over and silently no-ops, which is why nothing was highlighted
-  // until a click happened to re-run the paint against a settled DOM. We make
-  // it robust by (a) painting now, (b) painting on the next frame, and (c)
-  // watching the article subtree with a MutationObserver and repainting whenever
-  // react-markdown finally swaps in / replaces the rendered nodes. The CSS
-  // Custom Highlight API doesn't mutate the DOM, so this never loops.
-  useEffect(() => {
-    const el = articleRef.current;
-    if (!el) return;
-    const clear = editing || viewingSha;
-    const targets = clear
-      ? []
-      : commentThreads
-          .map((t) => t.root)
-          .filter(
-            // Resolved threads disappear from the panel, so drop their doc
-            // highlight too; orphaned ones have no live span to paint.
-            (r) =>
-              r.status !== "orphaned" &&
-              r.status !== "resolved" &&
-              r.start_offset !== null &&
-              r.end_offset !== null &&
-              r.quoted_text !== null,
-          )
-          .map((r) => ({
-            startOffset: r.start_offset as number,
-            endOffset: r.end_offset as number,
-            quotedText: r.quoted_text as string,
-            active: r.id === activeCommentId,
-          }));
-
-    let cancelled = false;
-    let raf = 0;
-    let attempts = 0;
-    const MAX_ATTEMPTS = 60; // ~1s at 60fps, then give up (some spans may be unmappable)
-    const tick = () => {
-      if (cancelled) return;
-      const painted = paintCommentHighlights(el, body, targets);
-      attempts += 1;
-      // The markdown text nodes commit a tick after React renders, so the first
-      // paint can find nothing to range over. Retry per-frame until every target
-      // lands (or we hit the cap) — this is what makes highlights appear on a
-      // fresh load instead of only after a click re-ran the paint.
-      if (painted < targets.length && attempts < MAX_ATTEMPTS) {
-        raf = requestAnimationFrame(tick);
-      }
-    };
-    tick();
-
-    // Repaint when react-markdown later swaps nodes (e.g. an edit/remap rerenders
-    // the body). Reset the retry budget so the fresh DOM gets a full set of tries.
-    const observer = new MutationObserver(() => {
-      attempts = 0;
-      cancelAnimationFrame(raf);
-      tick();
-    });
-    observer.observe(el, {
-      childList: true,
-      subtree: true,
-      characterData: true,
-    });
-
-    return () => {
-      cancelled = true;
-      cancelAnimationFrame(raf);
-      observer.disconnect();
-    };
-    // `loading` gates whether <article> is mounted: when the body is primed by
-    // the SWR cache before `loading` flips, `body` doesn't change on mount, so
-    // without this dep the effect wouldn't re-run and the article would mount
-    // with no paint (highlights only appeared after a click). Re-run on mount.
-  }, [commentThreads, body, editing, viewingSha, activeCommentId, loading]);
-
-  // Select a thread (its span gets the orange highlight) and scroll the doc to
-  // bring that span into view. Only an explicit click runs this — keying a
-  // scroll off `activeCommentId` alone would also re-scroll on every comment
-  // refetch while a thread stays selected. Deferred a frame so the layout (and
-  // the active repaint) settles before we measure the range.
-  const activateComment = useCallback(
-    (id: string | null) => {
-      setActiveCommentId(id);
-      if (!id || editing || viewingSha) return;
-      const el = articleRef.current;
-      if (!el) return;
-      const root = commentThreads.find((t) => t.root.id === id)?.root;
-      if (
-        !root ||
-        root.status === "orphaned" ||
-        root.status === "resolved" ||
-        root.start_offset === null ||
-        root.end_offset === null
-      )
-        return;
-      requestAnimationFrame(() => {
-        scrollCommentIntoView(el, body, root.start_offset!, root.end_offset!);
-      });
-    },
-    [commentThreads, body, editing, viewingSha],
-  );
-
-  // Deep-link: `?comment=<id>` opens the panel, selects that thread, and scrolls
-  // to its anchored span — the shareable-link counterpart to click-to-focus.
-  // Runs once per id (focusedCommentRef), and only once the thread is loaded.
-  // The scroll retries per-frame because on a fresh load react-markdown commits
-  // its text nodes a tick after this effect runs (same reason the highlight
-  // paint retries).
-  useEffect(() => {
-    const target = searchParams?.get("comment");
-    if (!target || loading || editing || viewingSha) return;
-    if (focusedCommentRef.current === target) return;
-    const root = commentThreads.find((t) => t.root.id === target)?.root;
-    if (!root) return; // not loaded yet, or not a thread on this page
-    focusedCommentRef.current = target;
-    setActiveCommentId(target);
-    openComments();
-    if (
-      root.status === "orphaned" ||
-      root.status === "resolved" ||
-      root.start_offset === null ||
-      root.end_offset === null
-    )
-      return; // selected, but no live span to scroll to
-    const el = articleRef.current;
-    if (!el) return;
-    let raf = 0;
-    let attempts = 0;
-    const tick = () => {
-      if (scrollCommentIntoView(el, body, root.start_offset!, root.end_offset!))
-        return;
-      if (++attempts < 60) raf = requestAnimationFrame(tick);
-    };
-    tick();
-    return () => cancelAnimationFrame(raf);
-  }, [
-    searchParams,
-    commentThreads,
-    loading,
-    editing,
-    viewingSha,
-    body,
-    openComments,
-  ]);
-
-  // On a text selection in the rendered article, offer a floating "Comment"
-  // affordance anchored above the selection (render mode only).
-  const onArticleMouseUp = useCallback(() => {
-    const el = articleRef.current;
-    const sel = window.getSelection();
-    if (!el || !sel || sel.rangeCount === 0) {
-      setSelTool(null);
-      return;
-    }
-    const draft = selectionToAnchor(el, body);
-    if (!draft) {
-      setSelTool(null);
-      return;
-    }
-    const rect = sel.getRangeAt(0).getBoundingClientRect();
-    setSelTool({ x: rect.left + rect.width / 2, y: rect.top, draft });
-  }, [body]);
-
-  useEffect(() => {
-    const onSel = () => {
-      const s = window.getSelection();
-      if (!s || s.isCollapsed) setSelTool(null);
-    };
-    document.addEventListener("selectionchange", onSel);
-    return () => document.removeEventListener("selectionchange", onSel);
-  }, []);
-
-  // Active-agents panel (collapsible chip near the top of the doc).
-  // We always know the count (so the chip can label "Active agents (N)"),
-  // but the entry list only renders when the user expands it.
-  const [agentsOpen, setAgentsOpen] = useState(false);
-  const [agents, setAgents] = useState<DocumentActivity[]>([]);
-  const [agentsError, setAgentsError] = useState<string | null>(null);
-  // Inline template gallery state. We show clickable cards above the
-  // editor whenever the draft is "empty enough" — either truly blank
-  // or still matching the body of the template the user just applied
-  // (so they can keep swapping without losing work). Once they edit on
-  // top of a template, the gallery disappears.
-  const [templates, setTemplates] = useState<DocumentTemplateSummary[] | null>(
-    null,
-  );
-  const [appliedTemplateBody, setAppliedTemplateBody] = useState<string | null>(
-    null,
-  );
-  const [appliedTemplateId, setAppliedTemplateId] = useState<string | null>(
-    null,
-  );
-  const [applyingTemplateId, setApplyingTemplateId] = useState<string | null>(
-    null,
-  );
-  const loadLatest = useCallback(() => {
-    setLoading(true);
-    setError(null);
-    setEditing(false);
-    setViewingSha(null);
-    setDiffData(null);
-    apiFetch<FileResponse>(`/wiki/file?path=${encodeURIComponent(path)}`)
-      .then((r) => {
-        setBody(r.body);
-        setHeadSha(r.head_sha ?? null);
-      })
-      .catch((e) => setError(e instanceof Error ? e.message : "failed to load"))
-      .finally(() => setLoading(false));
-  }, [path]);
-
-  useEffect(() => {
-    loadLatest();
-    setHistoryOpen(false);
-    setCommits(null);
-  }, [loadLatest]);
-
-  // SWR revalidates the doc so MCP-driven edit_doc writes appear
-  // without manual refresh. Skip when the user is editing the
-  // textarea or viewing an old commit — both would get clobbered.
-  const liveKey =
-    !editing && viewingSha === null
-      ? `/wiki/file?path=${encodeURIComponent(path)}`
-      : null;
-  const { data: liveDoc } = useSWR<FileResponse>(liveKey, {
-    refreshInterval: 1500,
-    revalidateOnFocus: true,
-    dedupingInterval: 0,
-  });
-
-  // Active external agent sessions on this page — surfaced in the
-  // Active agents bar alongside read/write activity.
-  const { sessions: agentSessions, refresh: refreshSessions } =
-    useAgentSessions(path);
-  const activeSessions = agentSessions.filter(
-    (s) =>
-      s.status === "active" ||
-      s.status === "idle" ||
-      // Onyx Craft (in_app) lifecycle states worth surfacing on the page.
-      (s.tool_id === "onyx-craft" &&
-        (s.status === "provisioning" ||
-          s.status === "ready" ||
-          s.status === "failed")),
-  );
-
-  const handleCloseSession = useCallback(
-    async (id: string) => {
-      if (
-        !(await confirmDialog({
-          title: "Close this agent session?",
-          confirmLabel: "Close session",
-        }))
-      )
-        return;
-      try {
-        await closeSession(id, "user_clicked");
-      } catch (err) {
-        alert(err instanceof Error ? err.message : "Failed to close session");
-      } finally {
-        await refreshSessions();
-      }
-    },
-    [refreshSessions, confirmDialog],
-  );
-  useEffect(() => {
-    if (!liveDoc) return;
-    // Re-check the gate at apply time — an in-flight fetch from before
-    // the user toggled `editing` can still resolve here and would
-    // otherwise clobber the live session buffer. While editing, the
-    // co-edit SSE stream is the source of truth, not this poll.
-    if (editing || viewingSha !== null) return;
-    // The read is session-aware: while a co-edit session is open, /wiki/file
-    // serves the live buffer (a quick HEAD+buffer merge), so this poll reflects
-    // saved edits immediately — no dependency on the async checkpoint commit.
-    setBody((prev) => (prev === liveDoc.body ? prev : liveDoc.body));
-    setHeadSha(liveDoc.head_sha ?? null);
-  }, [liveDoc, editing, viewingSha]);
-
-  // Fetch template summaries once; the gallery uses them as its menu
-  // and falls back to "no templates configured" when the list is empty.
-  useEffect(() => {
-    let cancelled = false;
-    listTemplateSummaries()
-      .then((rows) => {
-        if (!cancelled) setTemplates(rows);
-      })
-      .catch(() => {
-        if (!cancelled) setTemplates([]);
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, []);
-
-  // ``?new=1`` is set by NewDocView after first-create. The doc body
-  // was already chosen there, so we land in the rendered view (not
-  // edit mode); the param's only remaining job is to expand the chat
-  // widget so the assistant is right there on the freshly-created doc.
-  const isFreshDoc = searchParams?.get("new") === "1";
-
-  // Drafting state: per-doc template-seeded draft tracked server-side.
-  // We re-fetch when the path changes and after each save (the server
-  // clears the row once the body diverges from the template snapshot).
-  // ``?new=1`` triggers a one-shot expand of the chat widget so the
-  // assistant is immediately available while drafting.
-  const refreshDraftState = useCallback(async () => {
-    try {
-      const state = await getDraftState(path);
-      setDrafting(
-        state
-          ? {
-              kind: "template",
-              path: state.path,
-              templateName: state.template_name,
-              templateId: state.template_id,
-            }
-          : null,
-      );
-      return state;
-    } catch {
-      setDrafting(null);
-      return null;
-    }
-  }, [path, setDrafting]);
-
-  useEffect(() => {
-    let cancelled = false;
-    void (async () => {
-      await refreshDraftState();
-      if (!cancelled && isFreshDoc) requestExpand();
-    })();
-    return () => {
-      cancelled = true;
-    };
-  }, [refreshDraftState, isFreshDoc, requestExpand]);
-
-  // Clear drafting context on unmount so the banner doesn't follow the
-  // user to other pages.
-  useEffect(() => {
-    return () => setDrafting(null);
-  }, [setDrafting]);
-
-  const refreshAgents = useCallback(() => {
-    setAgentsError(null);
-    apiFetch<DocumentActivityResponse>(
-      `/wiki/file/activity?path=${encodeURIComponent(path)}`,
-    )
-      .then((r) => setAgents(r.agents))
-      .catch((e) =>
-        setAgentsError(
-          e instanceof Error ? e.message : "failed to load activity",
-        ),
-      );
-  }, [path]);
-
-  useEffect(() => {
-    refreshAgents();
-    setAgentsOpen(false);
-    // Refresh on window focus so the chip count tracks reality after
-    // the user comes back from another tab. The endpoint is cheap.
-    function onFocus() {
-      refreshAgents();
-    }
-    window.addEventListener("focus", onFocus);
-    return () => window.removeEventListener("focus", onFocus);
-  }, [refreshAgents]);
-
-  const refreshHistory = useCallback(async () => {
-    setHistoryError(null);
-    try {
-      const r = await fetchFileHistory(path);
-      setCommits(r.commits);
-      setHeadSha(r.head_sha);
-      return r;
-    } catch (e) {
-      setHistoryError(
-        e instanceof Error ? e.message : "failed to load history",
-      );
-      return null;
-    }
-  }, [path]);
-
-  function closeHistory() {
-    setHistoryOpen(false);
-    // Closing the panel exits history mode entirely — back to the rendered
-    // latest version. Skip while editing: `viewingSha` is the fork base the
-    // save needs, and the textarea is showing the user's draft anyway.
-    if (!editing && viewingSha !== null) {
-      setViewingSha(null);
-      setDiffData(null);
-    }
-  }
-
-  async function toggleHistory() {
-    if (historyOpen) {
-      closeHistory();
-      return;
-    }
-    setHistoryOpen(true);
-    // Mutual exclusion with the comments + policy + automations panels and
-    // the docked trigger editor (see ``openComments``).
-    setCommentsOpen(false);
-    setPolicyOpen(false);
-    setAutomationsOpen(false);
-    setTriggerModalOpen(false);
-    setEditingTrigger(null);
-    setCommentDraft(null);
-    // Opening history: show the newest commit's diff immediately rather
-    // than leaving the rendered body up until the user clicks a row.
-    const loaded = commits ?? (await refreshHistory())?.commits ?? null;
-    const newest = loaded?.[0];
-    if (newest && viewingSha === null) {
-      void onPickCommit(newest.sha);
-    }
-  }
-
-  async function onPickCommit(sha: string) {
-    if (sha === viewingSha) return;
-    setLoading(true);
-    setError(null);
-    setEditing(false);
-    try {
-      const r = await fetchFileDiff(path, sha);
-      setDiffData(r);
-      setViewingSha(sha);
-    } catch (e) {
-      setError(e instanceof Error ? e.message : "failed to load version");
-    } finally {
-      setLoading(false);
-    }
-  }
-
-  // Live co-editing session: clicking Edit joins the page's session and binds
-  // the textarea to the shared buffer; Save checkpoints, Cancel discards, a
-  // disconnect leaves. Only the current version is co-edited (the Edit button is
-  // hidden while viewing an old commit).
-  const coedit = useCoeditSession({
-    path,
-    enabled: editing,
-    committedBody: body,
-    myUserId: user?.id ?? null,
-    onEnd: () => {
-      setEditing(false);
-      setViewingSha(null);
-      // Don't refetch the body here. onSave sets it optimistically; the SWR
-      // poll then reconciles against the session-aware read (which serves the
-      // live buffer while the session is open), so there's no stale flash and
-      // no wait on the async checkpoint commit.
-      void refreshComments();
-      void refreshDraftState();
-    },
-  });
-
-  const segments = path.split("/");
-  const parentSlug = segments.slice(0, -1).join("/");
-  const currentBasename = segments[segments.length - 1] ?? path;
-  const currentBasenameNoExt = currentBasename.replace(/\.md$/i, "");
-  const trimmedFilename = filenameDraft.trim().replace(/^\/+|\/+$/g, "");
-  const filenameNoExt = trimmedFilename.replace(/\.md$/i, "");
-  const filenameValid = !!filenameNoExt && !filenameNoExt.includes("/");
-  const renamed =
-    editing && filenameValid && filenameNoExt !== currentBasenameNoExt;
-  // `viewingVersion`: a history version is displayed in the main pane.
-  // `viewingOld`: that version is not the newest commit for this file
-  // (`headSha` tracks `commits[0]`), so the warning banner applies.
-  const viewingVersion = viewingSha !== null;
-  const viewingOld = viewingVersion && viewingSha !== headSha;
-  // No navigate-away guard: editing writes to the shared session buffer, which
-  // is durable in Postgres and commits at session close — leaving the page
-  // never loses edits, so there's nothing to warn about.
-
-  function startEdit() {
-    // Entering edit mode joins the page's co-edit session (see the
-    // useCoeditSession hook, gated on `editing`), which seeds the buffer from
-    // current HEAD. The Edit button is hidden while viewing an old commit, so
-    // editing always targets the current version.
-    setFilenameDraft(currentBasenameNoExt);
-    setError(null);
-    setEditing(true);
-  }
-
-  async function onDone() {
-    if (!filenameValid) {
-      setError("Filename cannot be empty or contain '/'.");
-      return;
-    }
-    setSaving(true);
-    setError(null);
-    // Snapshot what we're committing so the viewer shows it immediately; the
-    // checkpoint commit runs on a worker, but the session-aware read serves the
-    // live buffer meanwhile, so the SWR poll keeps showing this content.
-    const finalText = coedit.buffer;
-    try {
-      // Commit the session buffer to git (checkpoint), then leave. onEnd exits
-      // edit mode + refreshes comments/drafting; co-editing removes the
-      // human-vs-human conflict, so there's no 409 path here. On a rename-only
-      // save the buffer is unchanged, so the checkpoint is a no-op server-side
-      // (version == checkpointed_version) — no empty commit before the move.
-      await coedit.save();
-      setBody(finalText);
-      setDiffData(null);
-      if (historyOpen) refreshHistory();
-      else setCommits(null);
-      if (renamed) {
-        const finalName = filenameNoExt + ".md";
-        const newRel = parentSlug ? `${parentSlug}/${finalName}` : finalName;
-        await apiFetch("/wiki/move", {
-          method: "POST",
-          body: JSON.stringify({ old_path: path, new_path: newRel }),
-        });
-        // The id URL survives a rename, so the open page's id→path resolve now
-        // points at the old path. Revalidate every wiki cache (including that
-        // resolve and the content read) so the page follows to its new path,
-        // then route to the renamed doc's durable id URL — falling back to the
-        // path URL only for an id-less page.
-        await revalidateWiki();
-        const id = (await resolveIds([newRel]))[newRel];
-        router.replace(id ? wikiHref(id) : `/app/wiki/${newRel}`);
-      }
-    } catch (e) {
-      setError(e instanceof Error ? e.message : "save failed");
-    } finally {
-      setSaving(false);
-    }
-  }
-
-  async function onPickTemplate(template: DocumentTemplateSummary) {
-    setApplyingTemplateId(template.id);
-    setError(null);
-    try {
-      const full = await getTemplate(template.id);
-      coedit.setDoc(full.body);
-      setAppliedTemplateBody(full.body);
-      setAppliedTemplateId(template.id);
-      await setDraftTemplate(path, template.id);
-      await refreshDraftState();
-    } catch (e) {
-      setError(e instanceof Error ? e.message : "failed to apply template");
-    } finally {
-      setApplyingTemplateId(null);
-    }
-  }
-
-  async function onPickBlank() {
-    coedit.setDoc("");
-    setAppliedTemplateBody(null);
-    setAppliedTemplateId(null);
-    setError(null);
-    try {
-      await setDraftTemplate(path, null);
-      // Don't go through ``refreshDraftState`` — the server only knows
-      // about template-backed drafts (``document_drafts``), so it would
-      // return null and clear drafting. Set blank-drafting locally
-      // instead so the chat widget spins up a generic kickoff session.
-      setDrafting({ kind: "blank", path });
-    } catch (e) {
-      setError(e instanceof Error ? e.message : "failed to clear template");
-    }
-  }
-
-  // Page actions live in the single pinned header (WikiHeader), not a second
-  // header inside the scroll area — they portal into its right-aligned slot.
-  // The panel toggles are icon SelectButtons so the open panel shows the
-  // selected tint; the others are tertiary icon Buttons with the lone solid
-  // CTA (Edit). Editing swaps in labelled Cancel / Save for clarity.
-  const headerActions = editing ? (
-    // One exit: "Done" leaves the session (committing the shared buffer). There
-    // is no Cancel — the buffer is the live shared document, so a single
-    // participant can't discard everyone's in-progress edits.
-    <Button variant="action" onClick={onDone} disabled={saving}>
-      {saving ? "Finishing…" : "Done"}
-    </Button>
-  ) : !loading && !error ? (
-    <>
-      <Button
-        icon={SvgSparkle}
-        prominence="tertiary"
-        tooltip="Run Agent"
-        onClick={() => setRunAgentOpen(true)}
-      />
-      <SelectButton
-        icon={SvgWorkflow}
-        state={automationsOpen ? "selected" : "empty"}
-        tooltip="Triggers"
-        onClick={() => {
-          if (automationsOpen || triggerModalOpen) {
-            setAutomationsOpen(false);
-            setTriggerModalOpen(false);
-            setEditingTrigger(null);
-            return;
-          }
-          setHistoryOpen(false);
-          setCommentsOpen(false);
-          setCommentDraft(null);
-          setPolicyOpen(false);
-          setAutomationsOpen(true);
-        }}
-      />
-      <Button
-        icon={SvgShare}
-        prominence="tertiary"
-        tooltip="Share"
-        onClick={() => setShareOpen(true)}
-      />
-      <SelectButton
-        icon={SvgHistory}
-        state={historyOpen ? "selected" : "empty"}
-        tooltip="History"
-        onClick={toggleHistory}
-      />
-      <SelectButton
-        icon={SvgBubbleText}
-        state={commentsOpen ? "selected" : "empty"}
-        tooltip="Comments"
-        onClick={() => (commentsOpen ? setCommentsOpen(false) : openComments())}
-      />
-      <SelectButton
-        icon={SvgShield}
-        state={policyOpen ? "selected" : "empty"}
-        tooltip="Update Policy"
-        onClick={() => {
-          if (policyOpen) {
-            setPolicyOpen(false);
-            return;
-          }
-          setHistoryOpen(false);
-          setCommentsOpen(false);
-          setCommentDraft(null);
-          setAutomationsOpen(false);
-          setTriggerModalOpen(false);
-          setEditingTrigger(null);
-          setPolicyOpen(true);
-        }}
-      />
-      {/* Editing always targets current HEAD via a co-edit session, so hide
-          Edit while viewing an older commit — "Back to latest" first. */}
-      {!viewingOld && (
-        <Button
-          icon={SvgEdit}
-          variant="action"
-          tooltip="Edit"
-          onClick={startEdit}
-        />
-      )}
-    </>
-  ) : null;
-
-  return (
-    <main
-      className={`box-border flex h-full min-h-0 flex-col ${isMobile ? "px-3 py-4" : "px-8 py-6"}`}
-    >
-      {host?.el && headerActions && createPortal(headerActions, host.el)}
-
-      {agentsBarHost?.el &&
-        !editing &&
-        createPortal(
-          <ActiveAgentsBar
-            agents={agents}
-            sessions={activeSessions}
-            error={agentsError}
-            open={agentsOpen}
-            onToggle={() => setAgentsOpen((v) => !v)}
-            onCloseSession={handleCloseSession}
-          />,
-          agentsBarHost.el,
-        )}
-
-      {!editing && <DocTitle path={path} />}
-
-      {!editing && triggerStatus && (
-        <div className="mb-3 text-xs text-(--text-04)">{triggerStatus}</div>
-      )}
-
-      <TriggerPanel
-        open={triggerModalOpen && (isMobile || !rightHost?.el)}
-        initial={editingTrigger ?? { scope_path: path }}
-        lockScope={!editingTrigger}
-        onDelete={
-          editingTrigger
-            ? async () => {
-                if (
-                  !(await confirmDialog({
-                    title: "Delete this trigger?",
-                    body: `"${editingTrigger.nl_description}"`,
-                    confirmLabel: "Delete",
-                  }))
-                )
-                  return;
-                await deleteTrigger(editingTrigger.id);
-                await refreshTriggers();
-                setTriggerModalOpen(false);
-                setEditingTrigger(null);
-              }
-            : undefined
-        }
-        onClose={() => {
-          setTriggerModalOpen(false);
-          setEditingTrigger(null);
-        }}
-        onSaved={(t) => {
-          setTriggerStatus(
-            editingTrigger
-              ? `Updated trigger for ${t.scope_path}`
-              : `Created trigger for ${t.scope_path}`,
-          );
-          void refreshTriggers();
-        }}
-      />
-
-      <ShareDialog
-        path={path}
-        open={shareOpen}
-        onClose={() => setShareOpen(false)}
-      />
-
-      <RunAgentPanel
-        open={runAgentOpen}
-        onClose={() => setRunAgentOpen(false)}
-        wikiPath={path || null}
-      />
-
-      {error && (
-        <div className="mb-3 rounded-(--border-radius-04) bg-(--status-error-01) p-[10px] text-[13px] text-(--status-text-error-05)">
-          {error}
-        </div>
-      )}
-
-      {viewingOld && !loading && !error && (
-        <div className="mb-3 flex items-center gap-3 rounded-(--border-radius-08) border border-(--status-warning-02) bg-(--status-warning-01) px-3 py-2 text-[13px] text-(--status-text-warning-05)">
-          <span>
-            Viewing an older version
-            {viewingSha ? ` (${viewingSha.slice(0, 7)})` : ""}.
-          </span>
-          <div className="flex-1" />
-          <Button size="sm" onClick={loadLatest}>
-            Back to latest
-          </Button>
-        </div>
-      )}
-
-      {loading && <LoadingSpinner />}
-
-      {!loading && !error && (
-        <>
-          {editing || (viewingVersion && diffData) ? (
-            // Editor / diff: fixed-height, capped + centered column with its own
-            // internal scroll (you edit against a pinned viewport, not a growing
-            // page).
-            <div className="flex min-h-0 flex-1 justify-center">
-              <div className="flex w-full max-w-[768px] min-w-0 flex-col gap-3">
-                {editing ? (
-                  <>
-                    <FilenameRow
-                      parent={parentSlug}
-                      value={filenameDraft}
-                      onChange={setFilenameDraft}
-                      disabled={saving}
-                    />
-                    <CoeditPresenceBar
-                      participants={coedit.participants}
-                      typing={coedit.typing}
-                      selfUserId={user?.id ?? null}
-                    />
-                    {(() => {
-                      // Cards visible while the body is still "empty enough"
-                      // to discard without losing user work: truly blank, or
-                      // still verbatim equal to the template the user just
-                      // applied (so they can keep swapping templates).
-                      const isBlank = coedit.buffer.trim() === "";
-                      const matchesApplied =
-                        appliedTemplateBody !== null &&
-                        coedit.buffer === appliedTemplateBody;
-                      const showGallery =
-                        (isBlank || matchesApplied) &&
-                        templates !== null &&
-                        templates.length > 0;
-                      if (!showGallery) return null;
-                      return (
-                        <TemplateGallery
-                          templates={templates!}
-                          activeId={matchesApplied ? appliedTemplateId : null}
-                          applyingId={applyingTemplateId}
-                          blankActive={isBlank && appliedTemplateId === null}
-                          onPick={(t) => void onPickTemplate(t)}
-                          onBlank={() => void onPickBlank()}
-                        />
-                      );
-                    })()}
-                    {coedit.session ? (
-                      <Coeditor
-                        key={coedit.session.id}
-                        session={coedit.session}
-                        peers={coedit.peers}
-                        onSelectionChange={coedit.reportSelection}
-                        onServerFrame={coedit.onServerFrame}
-                        reportDoc={coedit.reportDoc}
-                        registerFlush={coedit.registerFlush}
-                        registerSetDoc={coedit.registerSetDoc}
-                        placeholder="Start typing, or pick a template above…"
-                      />
-                    ) : (
-                      // Joining the session; the editor mounts once we have its
-                      // start version + doc.
-                      <div className="flex min-h-0 flex-1 items-center justify-center text-[13px] text-(--text-03)">
-                        Connecting…
-                      </div>
-                    )}
-                  </>
-                ) : (
-                  <div className="flex min-h-0 flex-1 overflow-hidden">
-                    <DiffView
-                      data={diffData!}
-                      commit={
-                        commits?.find((c) => c.sha === viewingSha) ?? undefined
-                      }
-                      loadBody={async () => {
-                        const sha = viewingSha;
-                        if (!sha) return "";
-                        const r = await apiFetch<FileResponse>(
-                          `/wiki/file?path=${encodeURIComponent(
-                            path,
-                          )}&ref=${encodeURIComponent(sha)}`,
-                        );
-                        return r.body;
-                      }}
-                    />
-                  </div>
-                )}
-              </div>
-            </div>
-          ) : (
-            // Render mode: the whole main-content area scrolls with the
-            // scrollbar flush at the far right edge (the negative margin cancels
-            // <main>'s horizontal padding, re-applied inside so the text keeps
-            // its gutter); the article text stays capped + centered.
-            <div
-              className={`min-h-0 flex-1 overflow-y-auto ${isMobile ? "-mx-3 px-3" : "-mx-8 px-8"}`}
-            >
-              <div className="mx-auto w-full max-w-[768px]">
-                <UpdateHealthBanner
-                  path={path}
-                  onOpenPolicy={() => {
-                    setHistoryOpen(false);
-                    setCommentsOpen(false);
-                    setAutomationsOpen(false);
-                    setTriggerModalOpen(false);
-                    setEditingTrigger(null);
-                    setPolicyOpen(true);
-                  }}
-                />
-              </div>
-              <article
-                ref={articleRef}
-                className="markdown mx-auto w-full max-w-[768px]"
-                onMouseUp={onArticleMouseUp}
-              >
-                {renderedBody}
-              </article>
-            </div>
-          )}
-          {/* Desktop side panels dock at the app's right edge (full height,
-              beside the header) by portaling into the shell's right-panel host,
-              so they never eat into the reading column above. Mobile keeps the
-              fixed sheet rendered below. */}
-          {historyOpen &&
-            !isMobile &&
-            rightHost?.el &&
-            createPortal(
-              <div className="flex h-full w-[400px] border-l border-(--border-01)">
-                <HistoryPanel
-                  commits={commits}
-                  error={historyError}
-                  headSha={headSha}
-                  viewingSha={viewingSha}
-                  onPick={onPickCommit}
-                  onClose={closeHistory}
-                  fullHeight
-                />
-              </div>,
-              rightHost.el,
-            )}
-          {commentsOpen &&
-            !isMobile &&
-            rightHost?.el &&
-            createPortal(
-              <div className="flex h-full w-[400px] border-l border-(--border-01)">
-                <CommentsPanel
-                  path={path}
-                  headSha={headSha}
-                  draft={commentDraft}
-                  threads={commentThreads}
-                  onChanged={refreshComments}
-                  activeId={activeCommentId}
-                  onActivate={activateComment}
-                  onDraftConsumed={() => setCommentDraft(null)}
-                  onClose={() => {
-                    setCommentsOpen(false);
-                    setCommentDraft(null);
-                  }}
-                  fullHeight
-                />
-              </div>,
-              rightHost.el,
-            )}
-          {policyOpen &&
-            !isMobile &&
-            rightHost?.el &&
-            createPortal(
-              <div className="flex h-full w-[400px] border-l border-(--border-01)">
-                <UpdatePolicyPanel
-                  path={path}
-                  onClose={() => setPolicyOpen(false)}
-                  onShowHistory={toggleHistory}
-                  fullHeight
-                />
-              </div>,
-              rightHost.el,
-            )}
-          {automationsOpen &&
-            !isMobile &&
-            rightHost?.el &&
-            createPortal(
-              <div className="flex h-full w-[480px] flex-col border-l border-(--border-01) bg-(--background-tint-01) p-2">
-                <TriggersSidePanel path={path} onStatus={setTriggerStatus} />
-              </div>,
-              rightHost.el,
-            )}
-        </>
-      )}
-      {historyOpen && isMobile && (
-        // Mobile: render history as a fixed slide-in sheet over the
-        // markdown content rather than a 320px side-panel that would
-        // squeeze the body to nothing on a 375px screen.
-        <>
-          <div
-            onClick={closeHistory}
-            aria-hidden
-            className="fixed inset-0 z-[60] bg-(--mask-03)"
-          />
-          <div className="fixed top-0 right-0 bottom-0 z-[70] flex w-[min(360px,100vw)] shadow-(--shadow-panel)">
-            <HistoryPanel
-              commits={commits}
-              error={historyError}
-              headSha={headSha}
-              viewingSha={viewingSha}
-              onPick={(sha) => {
-                onPickCommit(sha);
-                // Plain close (no reset): the sheet covers the content, so a
-                // deliberate pick must keep the chosen version visible. The
-                // banner's "Back to latest" is the way back.
-                setHistoryOpen(false);
-              }}
-              onClose={closeHistory}
-              fullHeight
-            />
-          </div>
-        </>
-      )}
-      {commentsOpen && isMobile && (
-        <>
-          <div
-            onClick={() => setCommentsOpen(false)}
-            aria-hidden
-            className="fixed inset-0 z-[60] bg-(--mask-03)"
-          />
-          <div className="fixed top-0 right-0 bottom-0 z-[70] flex w-[min(360px,100vw)] shadow-(--shadow-panel)">
-            <CommentsPanel
-              path={path}
-              headSha={headSha}
-              draft={commentDraft}
-              threads={commentThreads}
-              onChanged={refreshComments}
-              activeId={activeCommentId}
-              onActivate={activateComment}
-              onDraftConsumed={() => setCommentDraft(null)}
-              onClose={() => {
-                setCommentsOpen(false);
-                setCommentDraft(null);
-              }}
-              fullHeight
-            />
-          </div>
-        </>
-      )}
-      {policyOpen && isMobile && (
-        <>
-          <div
-            onClick={() => setPolicyOpen(false)}
-            aria-hidden
-            className="fixed inset-0 z-[60] bg-(--mask-03)"
-          />
-          <div className="fixed top-0 right-0 bottom-0 z-[70] flex w-[min(360px,100vw)] shadow-(--shadow-panel)">
-            <UpdatePolicyPanel
-              path={path}
-              onClose={() => setPolicyOpen(false)}
-              onShowHistory={toggleHistory}
-              fullHeight
-            />
-          </div>
-        </>
-      )}
-      {selTool && (
-        <div
-          onMouseDown={(e) => e.preventDefault()}
-          className="fixed z-[80] -translate-x-1/2 -translate-y-full rounded-(--border-radius-08) border border-(--border-01) bg-(--background-tint-01) p-1 shadow-(--shadow-popover)"
-          style={{
-            left: selTool.x,
-            top: selTool.y - 8,
-          }}
-        >
-          <Button
-            prominence="tertiary"
-            size="sm"
-            onClick={() => {
-              setCommentDraft(selTool.draft);
-              openComments();
-              setSelTool(null);
-              window.getSelection()?.removeAllRanges();
-            }}
-          >
-            💬 Comment
-          </Button>
-        </div>
-      )}
-    </main>
-  );
-}
-
-function ActiveAgentsBar({
-  agents,
-  sessions,
-  error,
-  open,
-  onToggle,
-  onCloseSession,
-}: {
-  agents: DocumentActivity[];
-  sessions: AgentSessionSummary[];
-  error: string | null;
-  open: boolean;
-  onToggle: () => void;
-  onCloseSession: (id: string) => void;
-}) {
-  const count = agents.length + sessions.length;
-  const expandable = count > 0;
-  return (
-    <div className="mb-3 overflow-hidden rounded-(--border-radius-08) border border-(--border-01) bg-(--background-tint-01)">
-      <button
-        onClick={expandable ? onToggle : undefined}
-        aria-expanded={expandable ? open : undefined}
-        disabled={!expandable}
-        className={`flex w-full items-center gap-2 border-none bg-transparent px-3 py-2 text-left text-[13px] ${expandable ? "cursor-pointer text-(--text-05)" : "cursor-default text-(--text-03)"}`}
-      >
-        <span
-          aria-hidden
-          className={`flex shrink-0 transition-transform duration-[120ms] ease-in-out ${open ? "rotate-90" : "rotate-0"} ${!expandable ? "text-(--text-02)" : "text-(--text-03)"}`}
-        >
-          <SvgChevronRight size={10} />
-        </span>
-        <span className="font-medium">
-          {expandable ? "Active agents" : "No agents active"}
-        </span>
-        {expandable && (
-          <span className="rounded-full bg-(--background-tint-03) px-[6px] py-[1px] text-[11px] font-semibold text-(--text-05)">
-            {count}
-          </span>
-        )}
-        {error && (
-          <span className="ml-auto text-xs text-(--status-text-error-05)">
-            {error}
-          </span>
-        )}
-      </button>
-      {expandable && open && (
-        <ul className="m-0 list-none border-t border-(--border-01) bg-(--background-tint-00) p-0">
-          {sessions.map((s, i) => (
-            <ActiveSessionRow
-              key={s.id}
-              s={s}
-              isLast={agents.length === 0 && i === sessions.length - 1}
-              onClose={() => onCloseSession(s.id)}
-            />
-          ))}
-          {agents.map((a, i) => (
-            <ActiveAgentRow
-              key={`${a.owner_display}-${a.agent_name ?? ""}-${
-                a.activity
-              }-${i}`}
-              a={a}
-              isLast={i === agents.length - 1}
-            />
-          ))}
-        </ul>
-      )}
-    </div>
-  );
-}
-
-function ActiveAgentRow({
-  a,
-  isLast,
-}: {
-  a: DocumentActivity;
-  isLast: boolean;
-}) {
-  return (
-    <li
-      className={
-        `flex items-center gap-[10px] overflow-hidden px-3 py-[10px] text-[13px] whitespace-nowrap` +
-        (isLast ? `` : ` border-b border-(--border-01)`)
-      }
-    >
-      <span className="shrink-0 rounded-(--border-radius-04) bg-(--background-tint-03) px-[6px] py-[1px] text-[10px] font-semibold tracking-[0.3px] text-(--text-05) uppercase">
-        {a.activity}
-      </span>
-
-      <span className="shrink-0 font-medium text-(--text-05)">
-        {a.owner_display}
-      </span>
-      {a.agent_name ? (
-        <span className="shrink-0 text-(--text-03)">
-          {"·"} {a.agent_name}
-        </span>
-      ) : null}
-
-      {a.description ? (
-        <span
-          className="min-w-0 grow overflow-hidden text-ellipsis text-(--text-04) italic"
-          title={a.description}
-        >
-          {"“"}
-          {a.description}
-          {"”"}
-        </span>
-      ) : (
-        <span className="flex-1" />
-      )}
-
-      <span
-        className="shrink-0 text-[11px] text-(--text-02)"
-        title={`Started ${absoluteTime(
-          a.registered_at,
-        )} · Expires ${absoluteTime(a.expires_at)}`}
-      >
-        {relativeTime(a.registered_at, "short")} {"·"} expires{" "}
-        {relativeTime(a.expires_at, "short")}
-      </span>
-    </li>
-  );
-}
-
-function ActiveSessionRow({
-  s,
-  isLast,
-  onClose,
-}: {
-  s: AgentSessionSummary;
-  isLast: boolean;
-  onClose: () => void;
-}) {
-  return (
-    <li
-      className={
-        `flex items-center gap-[10px] overflow-hidden px-3 py-[10px] text-[13px] whitespace-nowrap` +
-        (isLast ? `` : ` border-b border-(--border-01)`)
-      }
-    >
-      <span className="shrink-0 rounded-(--border-radius-04) bg-(--background-tint-03) px-[6px] py-[1px] text-[10px] font-semibold tracking-[0.3px] text-(--text-05) uppercase">
-        {s.status}
-      </span>
-
-      <span className="shrink-0 font-medium text-(--text-05)">{s.tool_id}</span>
-
-      {s.tool_id === "onyx-craft" && s.status === "failed" ? (
-        <span
-          className="min-w-0 grow overflow-hidden text-ellipsis text-(--status-text-error-05)"
-          title={craftFailureMessage(s.failure_reason)}
-        >
-          {craftFailureMessage(s.failure_reason)}
-        </span>
-      ) : (
-        <>
-          <span className="flex-1" />
-          <span
-            className="shrink-0 text-[11px] text-(--text-02)"
-            title={`Started ${absoluteTime(s.started_at)}`}
-          >
-            started {relativeTime(s.started_at, "short")}
-          </span>
-        </>
-      )}
-
-      {s.tool_id === "onyx-craft" && s.status === "ready" && s.external_url && (
-        <Button
-          type="button"
-          variant="default"
-          size="sm"
-          icon={SvgExternalLink}
-          onClick={() =>
-            window.open(
-              s.external_url as string,
-              "_blank",
-              "noopener,noreferrer",
-            )
-          }
-        >
-          Open Craft
-        </Button>
-      )}
-
-      <Button type="button" variant="default" size="sm" onClick={onClose}>
-        Close
-      </Button>
-    </li>
-  );
-}
-
-/** Every folder path in the tree (plus root ""), for the destination picker. */
-function collectFolders(entries: DocEntry[]): string[] {
-  const set = new Set<string>([""]);
-  for (const e of entries) {
-    const parts = e.path.split("/");
-    parts.pop();
-    let cur = "";
-    for (const p of parts) {
-      cur = cur ? `${cur}/${p}` : p;
-      set.add(cur);
-    }
-  }
-  return [...set].sort((a, b) => a.localeCompare(b));
-}
-
-/** Folder picker for the new-doc destination. "" is the wiki root ("Home"). */
-function DestinationSelect({
-  value,
-  folders,
-  onChange,
-  disabled,
-}: {
-  value: string;
-  folders: string[];
-  onChange: (v: string) => void;
-  disabled: boolean;
-}) {
-  const [open, setOpen] = useState(false);
-  return (
-    <Popover open={open} onOpenChange={setOpen}>
-      <Popover.Trigger asChild>
-        <span className="inline-flex max-w-full">
-          <OpenButton
-            variant="select-light"
-            size="md"
-            rounding="sm"
-            disabled={disabled}
-          >
-            {value === "" ? "Home" : value}
-          </OpenButton>
-        </span>
-      </Popover.Trigger>
-      <Popover.Content width="fit" align="start" sideOffset={4}>
-        <div className="max-h-[320px] max-w-[360px] min-w-[200px] overflow-y-auto">
-          <PopoverMenu>
-            {folders.map((f) => (
-              <LineItemButton
-                key={f || "__root__"}
-                icon={SvgFolder}
-                title={f === "" ? "Home" : f}
-                sizePreset="main-ui"
-                variant="body"
-                state={value === f ? "selected" : "empty"}
-                onClick={() => {
-                  onChange(f);
-                  setOpen(false);
-                }}
-              />
-            ))}
-          </PopoverMenu>
-        </div>
-      </Popover.Content>
-    </Popover>
-  );
-}
-
-// Co-editing presence while editing: who else is in the session and who's
-// typing right now. Remote carets aren't drawn in a plain textarea (that waits
-// for CodeMirror) — but edits already merge into the shared buffer live, so
-// this makes the collaboration visible. Renders nothing when you're alone.
-function CoeditPresenceBar({
-  participants,
-  typing,
-  selfUserId,
-}: {
-  participants: CoeditParticipant[];
-  typing: string[];
-  selfUserId: string | null;
-}) {
-  const others = participants.filter((p) => p.user_id !== selfUserId);
-  if (others.length === 0) return null;
-  const typingSet = new Set(typing);
-  return (
-    <div className="flex shrink-0 flex-wrap items-center gap-x-2 gap-y-1 text-[12px] text-(--text-03)">
-      <span
-        className="inline-block h-[7px] w-[7px] rounded-full bg-(--status-success-05)"
-        aria-hidden
-      />
-      {others.map((p) => (
-        <span key={p.user_id} className="inline-flex items-center gap-1">
-          <span className="font-medium text-(--text-04)">{p.user_display}</span>
-          {typingSet.has(p.user_id) && (
-            <span className="text-(--text-03) italic">typing…</span>
-          )}
-        </span>
-      ))}
-      <span>also editing</span>
-    </div>
-  );
-}
-
-function FilenameRow({
-  parent,
-  value,
-  onChange,
-  disabled,
-  autoFocus = false,
-  placeholder = "filename",
-}: {
-  parent: string;
-  value: string;
-  onChange: (v: string) => void;
-  disabled: boolean;
-  autoFocus?: boolean;
-  placeholder?: string;
-}) {
-  return (
-    <div className="flex shrink-0 items-stretch overflow-hidden rounded-(--border-radius-04) border border-(--border-01) bg-(--background-tint-00)">
-      {parent && (
-        <span className="flex items-center border-r border-(--border-01) bg-(--background-tint-02) px-[10px] font-mono text-[13px] text-(--text-04)">
-          {parent}/
-        </span>
-      )}
-      <input
-        // eslint-disable-next-line jsx-a11y/no-autofocus
-        autoFocus={autoFocus}
-        value={value.replace(/\.md$/i, "")}
-        onChange={(e) => onChange(e.target.value)}
-        placeholder={placeholder}
-        disabled={disabled}
-        spellCheck={false}
-        className="flex-1 border-none bg-transparent px-[10px] py-2 font-mono text-sm outline-none"
-      />
-      <span
-        aria-hidden
-        className="flex items-center border-l border-(--border-01) bg-(--background-tint-02) px-[10px] font-mono text-[13px] font-semibold text-(--text-04)"
-      >
-        .md
-      </span>
-    </div>
-  );
-}
+export { default } from "@/views/wiki/WikiPage";
